Refactor large files for better separation
