+++ __tests__/feedbackPreservation.test.ts
@@ -11,16 +11,28 @@ import { join } from 'path';
  */
 
 describe('Feedback preservation on team deletion', () => {
-  const serverSource = readFileSync(
-    join(__dirname, '..', 'server.js'),
+  const dataStoreSource = readFileSync(
+    join(__dirname, '..', 'server', 'services', 'dataStore.js'),
+    'utf-8'
+  );
+  const teamRoutesSource = readFileSync(
+    join(__dirname, '..', 'server', 'routes', 'teamRoutes.js'),
+    'utf-8'
+  );
+  const feedbackRoutesSource = readFileSync(
+    join(__dirname, '..', 'server', 'routes', 'feedbackRoutes.js'),
+    'utf-8'
+  );
+  const superAdminRoutesSource = readFileSync(
+    join(__dirname, '..', 'server', 'routes', 'superAdminRoutes.js'),
     'utf-8'
   );
 
   describe('Data normalization', () => {
     it('should initialize orphanedFeedbacks in normalizePersistedData', () => {
       // normalizePersistedData should set orphanedFeedbacks = [] if missing
-      expect(serverSource).toContain("normalized.orphanedFeedbacks");
-      expect(serverSource).toContain("Array.isArray(normalized.orphanedFeedbacks)");
+      expect(dataStoreSource).toContain("normalized.orphanedFeedbacks");
+      expect(dataStoreSource).toContain("Array.isArray(normalized.orphanedFeedbacks)");
     });
   });
 
@@ -28,9 +40,9 @@ describe('Feedback preservation on team deletion', () => {
     it('should preserve feedbacks before removing the team', () => {
       // The team deletion handler should copy feedbacks to orphanedFeedbacks
       // before splicing the team from the array
-      const deleteSection = serverSource.substring(
-        serverSource.indexOf("// POST /api/team/:teamId/delete"),
-        serverSource.indexOf("// GET /api/team/exists")
+      const deleteSection = teamRoutesSource.substring(
+        teamRoutesSource.indexOf("app.post('/api/team/:teamId/delete'"),
+        teamRoutesSource.indexOf("app.get('/api/team/exists")
       );
 
       // Should reference teamFeedbacks and orphanedFeedbacks
@@ -47,9 +59,9 @@ describe('Feedback preservation on team deletion', () => {
 
   describe('Feedback loading endpoints', () => {
     it('should include orphaned feedbacks in /api/feedbacks/all', () => {
-      const feedbacksAllSection = serverSource.substring(
-        serverSource.indexOf("// Get all feedbacks from all teams"),
-        serverSource.indexOf("// Add a comment to a feedback")
+      const feedbacksAllSection = feedbackRoutesSource.substring(
+        feedbackRoutesSource.indexOf("app.post('/api/feedbacks/all'"),
+        feedbackRoutesSource.indexOf("app.post('/api/feedbacks/comment'")
       );
 
       expect(feedbacksAllSection).toContain('orphanedFeedbacks');
@@ -57,9 +69,9 @@ describe('Feedback preservation on team deletion', () => {
     });
 
     it('should include orphaned feedbacks in /api/super-admin/feedbacks', () => {
-      const superAdminFeedbacksSection = serverSource.substring(
-        serverSource.indexOf("app.post('/api/super-admin/feedbacks'"),
-        serverSource.indexOf("app.post('/api/super-admin/update-email'")
+      const superAdminFeedbacksSection = superAdminRoutesSource.substring(
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/feedbacks'"),
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/update-email'")
       );
 
       expect(superAdminFeedbacksSection).toContain('orphanedFeedbacks');
@@ -69,45 +81,45 @@ describe('Feedback preservation on team deletion', () => {
 
   describe('Feedback operations on orphaned feedbacks', () => {
     it('should check orphaned feedbacks in team comment endpoint', () => {
-      const commentSection = serverSource.substring(
-        serverSource.indexOf("// Add a comment to a feedback"),
-        serverSource.indexOf("// Delete a comment from a feedback")
+      const commentSection = feedbackRoutesSource.substring(
+        feedbackRoutesSource.indexOf("app.post('/api/feedbacks/comment'"),
+        feedbackRoutesSource.indexOf("app.post('/api/feedbacks/comment/delete'")
       );
 
       expect(commentSection).toContain('orphanedFeedbacks');
     });
 
     it('should check orphaned feedbacks in team comment delete endpoint', () => {
-      const commentDeleteSection = serverSource.substring(
-        serverSource.indexOf("// Delete a comment from a feedback"),
-        serverSource.indexOf("// Delete a feedback (only the author")
+      const commentDeleteSection = feedbackRoutesSource.substring(
+        feedbackRoutesSource.indexOf("app.post('/api/feedbacks/comment/delete'"),
+        feedbackRoutesSource.indexOf("app.post('/api/feedbacks/delete'")
       );
 
       expect(commentDeleteSection).toContain('orphanedFeedbacks');
     });
 
     it('should check orphaned feedbacks in super-admin feedback update endpoint', () => {
-      const updateSection = serverSource.substring(
-        serverSource.indexOf("app.post('/api/super-admin/feedbacks/update'"),
-        serverSource.indexOf("app.post('/api/super-admin/feedbacks/delete'")
+      const updateSection = superAdminRoutesSource.substring(
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/feedbacks/update'"),
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/feedbacks/delete'")
       );
 
       expect(updateSection).toContain('orphanedFeedbacks');
     });
 
     it('should check orphaned feedbacks in super-admin feedback delete endpoint', () => {
-      const deleteSection = serverSource.substring(
-        serverSource.indexOf("app.post('/api/super-admin/feedbacks/delete'"),
-        serverSource.indexOf("// Super admin adds a comment to a feedback")
+      const deleteSection = superAdminRoutesSource.substring(
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/feedbacks/delete'"),
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/feedbacks/comment'")
       );
 
       expect(deleteSection).toContain('orphanedFeedbacks');
     });
 
     it('should check orphaned feedbacks in super-admin feedback comment endpoint', () => {
-      const commentSection = serverSource.substring(
-        serverSource.indexOf("// Super admin adds a comment to a feedback"),
-        serverSource.indexOf("// Send notification to team about new admin comment")
+      const commentSection = superAdminRoutesSource.substring(
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/feedbacks/comment'"),
+        superAdminRoutesSource.indexOf("app.post('/api/super-admin/update-password'")
       );
 
       expect(commentSection).toContain('orphanedFeedbacks');
@@ -116,14 +128,8 @@ describe('Feedback preservation on team deletion', () => {
 
   describe('Feedback preservation data flow', () => {
     it('should copy teamId and teamName from the deleted team', () => {
-      const deleteSection = serverSource.substring(
-        serverSource.indexOf("// Preserve feedbacks from the deleted team"),
-        serverSource.indexOf("currentData.teams.splice")
-      );
-
-      // Should map feedbacks with teamId and teamName
-      expect(deleteSection).toContain('teamId: f.teamId || deletedTeam.id');
-      expect(deleteSection).toContain('teamName: f.teamName || deletedTeam.name');
+      expect(teamRoutesSource).toContain('teamId: f.teamId || deletedTeam.id');
+      expect(teamRoutesSource).toContain('teamName: f.teamName || deletedTeam.name');
     });
   });
 });
