+++ package.json
@@ -37,7 +37,7 @@
     "security:audit": "npm audit --audit-level=moderate",
     "security:fix": "npm audit fix",
     "ci": "npm run lint && npm run type-check && npm run test && npm run build",
-    "prepare": "node -e \"try { require('husky') } catch (e) { if (e.code !== 'MODULE_NOT_FOUND') throw e }\""
+    "prepare": "node -e \"import('husky').catch((e) => { if (e.code !== 'ERR_MODULE_NOT_FOUND' && e.code !== 'MODULE_NOT_FOUND') throw e; })\""
   },
   "lint-staged": {
     "*.{ts,tsx}": [
