+++ server.js
@@ -3,18 +3,24 @@ import { createServer } from 'http';
 import { Server } from 'socket.io';
 import { fileURLToPath } from 'url';
 import { dirname, join } from 'path';
-import fs from 'fs';
-import { gzipSync, gunzipSync } from 'zlib';
-import nodemailer from 'nodemailer';
-import Database from 'better-sqlite3';
-import pg from 'pg';
-import { createHash, randomBytes, timingSafeEqual } from 'crypto';
-import rateLimit, { ipKeyGenerator } from 'express-rate-limit';
-import { createAdapter as createRedisAdapter } from '@socket.io/redis-adapter';
-import { createAdapter as createPostgresAdapter } from '@socket.io/postgres-adapter';
-import { createClient } from 'redis';
-import { resolveSocketAdapterStrategy, SOCKET_ADAPTER_STRATEGIES } from './socketAdapter.js';
-import { compactInviteLink } from './utils/inviteLink.js';
+
+import { createRateLimiters } from './server/config/rateLimiters.js';
+import { createDataStore } from './server/services/dataStore.js';
+import { createLogService } from './server/services/logService.js';
+import { createMailerService } from './server/services/mailerService.js';
+import { createTeamService } from './server/services/teamService.js';
+import { createTokenService } from './server/services/sessionTokens.js';
+import { createVersionService } from './server/services/versionService.js';
+import { initSocketAdapter } from './server/services/socketAdapter.js';
+import { registerSocketHandlers } from './server/services/socketHandlers.js';
+import { escapeHtml, sanitizeEmailLink, secureCompare, hashResetToken, pruneResetTokens } from './server/services/security.js';
+
+import { registerCoreRoutes } from './server/routes/coreRoutes.js';
+import { registerFeedbackRoutes } from './server/routes/feedbackRoutes.js';
+import { registerPasswordResetRoutes } from './server/routes/passwordResetRoutes.js';
+import { registerPublicRoutes } from './server/routes/publicRoutes.js';
+import { registerSuperAdminRoutes } from './server/routes/superAdminRoutes.js';
+import { registerTeamRoutes } from './server/routes/teamRoutes.js';
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -34,3222 +40,118 @@ const trustProxySetting = (() => {
 
 app.set('trust proxy', trustProxySetting);
 
-// CORS configuration - defaults to permissive for self-hosted deployments
-// Set CORS_ORIGIN to restrict in production (e.g., "https://retro.example.com")
-const corsOrigin = process.env.CORS_ORIGIN || "*";
+const corsOrigin = process.env.CORS_ORIGIN || '*';
 
 const io = new Server(server, {
-  // Explicit path avoids collisions with platform proxies
   path: '/socket.io',
   cors: {
     origin: corsOrigin,
-    methods: ["GET", "POST"]
+    methods: ['GET', 'POST']
   }
 });
 
-const buildRedisConfig = () => {
-  if (process.env.REDIS_URL) {
-    return { url: process.env.REDIS_URL };
-  }
-
-  const host = process.env.REDIS_HOST;
-  const port = Number(process.env.REDIS_PORT || 6379);
-  const password = process.env.REDIS_PASSWORD;
-
-  if (host) {
-    return {
-      socket: { host, port },
-      password: password || undefined
-    };
-  }
-
-  return null;
-};
-
-const initRedisAdapter = async (redisConfig) => {
-  try {
-    const pubClient = createClient(redisConfig);
-    const subClient = pubClient.duplicate();
-
-    await Promise.all([pubClient.connect(), subClient.connect()]);
-    io.adapter(createRedisAdapter(pubClient, subClient));
-    console.info('[Server] Using Redis adapter for Socket IO (multi-pod ready)');
-    return true;
-  } catch (err) {
-    console.error('[Server] Failed to initialize Redis adapter', err);
-    return false;
-  }
-};
-
-const initPostgresAdapter = async () => {
-  if (!pgPool) {
-    return false;
-  }
-
-  try {
-    await pgPool.query(`
-      CREATE TABLE IF NOT EXISTS socket_io_attachments (
-        id BIGSERIAL UNIQUE,
-        created_at TIMESTAMPTZ DEFAULT NOW(),
-        payload BYTEA
-      )
-    `);
-
-    io.adapter(createPostgresAdapter(pgPool));
-    console.info('[Server] Using PostgreSQL adapter for Socket IO (multi-pod ready)');
-    return true;
-  } catch (err) {
-    console.error('[Server] Failed to initialize PostgreSQL adapter', err);
-    return false;
-  }
-};
-
-const initSocketAdapter = async () => {
-  const redisConfig = buildRedisConfig();
-  const strategy = resolveSocketAdapterStrategy({
-    hasRedisConfig: !!redisConfig,
-    usePostgres
-  });
-
-  if (strategy === SOCKET_ADAPTER_STRATEGIES.REDIS) {
-    return initRedisAdapter(redisConfig);
-  }
-
-  if (strategy === SOCKET_ADAPTER_STRATEGIES.POSTGRES) {
-    return initPostgresAdapter();
-  }
-
-  console.info('[Server] Using in-memory Socket IO adapter (single-pod)');
-  return false;
-};
-
-// ==================== SESSION TOKEN MANAGEMENT ====================
-// Secure session tokens for browser refresh persistence (not stored in localStorage as password)
-
-const SESSION_TOKEN_EXPIRY_MS = 7 * 24 * 60 * 60 * 1000; // 7 days
-const sessionTokens = new Map(); // token -> { teamId, visitorId, createdAt }
-
-const generateSessionToken = () => {
-  return randomBytes(32).toString('hex');
-};
-
-const createSessionToken = (teamId, visitorId) => {
-  const token = generateSessionToken();
-  sessionTokens.set(token, {
-    teamId,
-    visitorId,
-    createdAt: Date.now()
-  });
-  return token;
-};
-
-const validateSessionToken = (token) => {
-  const session = sessionTokens.get(token);
-  if (!session) return null;
-
-  // Check expiry
-  if (Date.now() - session.createdAt > SESSION_TOKEN_EXPIRY_MS) {
-    sessionTokens.delete(token);
-    return null;
-  }
-
-  return session;
-};
-
-const invalidateSessionToken = (token) => {
-  sessionTokens.delete(token);
-};
-
-// Cleanup expired tokens periodically (every hour)
-setInterval(() => {
-  const now = Date.now();
-  for (const [token, session] of sessionTokens.entries()) {
-    if (now - session.createdAt > SESSION_TOKEN_EXPIRY_MS) {
-      sessionTokens.delete(token);
-    }
-  }
-  // Also cleanup expired super admin tokens
-  for (const [token, session] of superAdminTokens.entries()) {
-    if (now - session.createdAt > SESSION_TOKEN_EXPIRY_MS) {
-      superAdminTokens.delete(token);
-    }
-  }
-}, 60 * 60 * 1000);
-
-// Super Admin session tokens (separate from team session tokens)
-const superAdminTokens = new Map(); // token -> { createdAt }
-
-const createSuperAdminToken = () => {
-  const token = generateSessionToken();
-  superAdminTokens.set(token, { createdAt: Date.now() });
-  return token;
-};
-
-const validateSuperAdminToken = (token) => {
-  const session = superAdminTokens.get(token);
-  if (!session) return false;
-
-  if (Date.now() - session.createdAt > SESSION_TOKEN_EXPIRY_MS) {
-    superAdminTokens.delete(token);
-    return false;
-  }
-
-  return true;
-};
-
-/**
- * Validate super admin authentication using either password or session token
- * Returns true if authenticated, false otherwise
- */
-const validateSuperAdminAuth = (body) => {
-  if (!SUPER_ADMIN_PASSWORD) return false;
-
-  const { password, sessionToken } = body || {};
-
-  // Check password first
-  if (password && secureCompare(password, SUPER_ADMIN_PASSWORD)) {
-    return true;
-  }
-
-  // Fall back to session token
-  if (sessionToken && validateSuperAdminToken(sessionToken)) {
-    return true;
-  }
-
-  return false;
-};
-
-const escapeHtml = (value = '') => {
-  return String(value).replace(/[&<>"']/g, (char) => {
-    switch (char) {
-      case '&':
-        return '&amp;';
-      case '<':
-        return '&lt;';
-      case '>':
-        return '&gt;';
-      case '"':
-        return '&quot;';
-      case "'":
-        return '&#39;';
-      default:
-        return char;
-    }
-  });
-};
-
-const sanitizeEmailLink = (value) => {
-  if (!value) {
-    return '';
-  }
-
-  try {
-    const url = new URL(String(value));
-    if (url.protocol !== 'http:' && url.protocol !== 'https:') {
-      return '';
-    }
-    return url.toString();
-  } catch {
-    return '';
-  }
-};
-
-// Health endpoints for platform monitoring
-app.get('/health', (_req, res) => res.status(200).send('OK'));
-app.get('/ready', (_req, res) => res.status(200).send('READY'));
-
-// ==================== VERSION & ANNOUNCEMENTS ====================
-
-/**
- * Parse CHANGELOG.md and extract version announcements
- * @returns {{ current: string, announcements: Array<{ version: string, date: string, items: Array<{ type: string, description: string }> }> }}
- */
-const parseVersionAndChangelog = () => {
-  let currentVersion = '1.0';
-  const announcements = [];
-
-  // Read VERSION file
-  try {
-    const versionPath = join(__dirname, 'VERSION');
-    if (fs.existsSync(versionPath)) {
-      currentVersion = fs.readFileSync(versionPath, 'utf8').trim();
-    }
-  } catch (err) {
-    console.warn('[Server] Failed to read VERSION file:', err?.message);
-  }
-
-  // Parse CHANGELOG.md
-  try {
-    const changelogPath = join(__dirname, 'CHANGELOG.md');
-    if (fs.existsSync(changelogPath)) {
-      const content = fs.readFileSync(changelogPath, 'utf8');
-
-      // Split content by version headers
-      const versionBlocks = content.split(/(?=^## \[)/m).filter(block => block.trim());
-
-      for (const block of versionBlocks) {
-        const headerMatch = block.match(/^## \[([^\]]+)\] - (\d{4}-\d{2}-\d{2})/);
-        if (!headerMatch) continue;
-
-        const version = headerMatch[1];
-        const date = headerMatch[2];
-        const items = [];
-
-        // Map section names to announcement types
-        const typeMap = {
-          'Added': 'feature',
-          'Changed': 'improvement',
-          'Fixed': 'fix',
-          'Removed': 'removed',
-          'Security': 'security'
-        };
-
-        // Find all sections and their items
-        const sections = block.split(/^### /m).slice(1);
-        for (const section of sections) {
-          const lines = section.split('\n');
-          const sectionName = lines[0].trim();
-          const type = typeMap[sectionName];
-
-          if (!type) continue;
-
-          // Extract bullet points (lines starting with -)
-          for (let i = 1; i < lines.length; i++) {
-            const line = lines[i].trim();
-            if (line.startsWith('-') && !line.match(/^-+$/)) {
-              const description = line.substring(1).trim();
-              if (description && !description.startsWith('<!--') && !description.match(/^-+$/)) {
-                items.push({ type, description });
-              }
-            }
-          }
-        }
-
-        if (items.length > 0) {
-          announcements.push({ version, date, items });
-        }
-      }
-    }
-  } catch (err) {
-    console.warn('[Server] Failed to parse CHANGELOG.md:', err?.message);
-  }
-
-  return { current: currentVersion, announcements };
-};
-
-// Cache version info (reload on each request in dev, cache in production)
-let cachedVersionInfo = null;
-let versionCacheTime = 0;
-const VERSION_CACHE_TTL = process.env.NODE_ENV === 'production' ? 60000 : 0; // 1 minute in prod
-
-app.get('/api/version', (_req, res) => {
-  const now = Date.now();
-  if (!cachedVersionInfo || (now - versionCacheTime) > VERSION_CACHE_TTL) {
-    cachedVersionInfo = parseVersionAndChangelog();
-    versionCacheTime = now;
-  }
-  res.json(cachedVersionInfo);
-});
-
-app.use(express.json({ limit: '1mb' }));
+const SUPER_ADMIN_PASSWORD = process.env.SUPER_ADMIN_PASSWORD;
 
 const shouldSkipSuperAdminLimit = (req) => {
   if (req.path.startsWith('/api/super-admin')) {
-    return !process.env.SUPER_ADMIN_PASSWORD;
+    return !SUPER_ADMIN_PASSWORD;
   }
   return false;
 };
 
-// Rate limiting for authentication endpoints to prevent brute force attacks
-const authLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 5, // Limit each IP to 5 requests per windowMs
-  message: { error: 'too_many_attempts', retryAfter: '15 minutes' },
-  standardHeaders: true, // Return rate limit info in the `RateLimit-*` headers
-  legacyHeaders: false, // Disable the `X-RateLimit-*` headers
-  // Skip rate limiting if super admin is not configured
-  skip: shouldSkipSuperAdminLimit
+const dataStore = createDataStore({ rootDir: __dirname });
+const versionService = createVersionService({ rootDir: __dirname });
+const mailerService = createMailerService();
+const logService = createLogService();
+const tokenService = createTokenService({
+  secureCompare,
+  superAdminPassword: SUPER_ADMIN_PASSWORD
+});
+const teamService = createTeamService({ dataStore });
+const rateLimiters = createRateLimiters({ shouldSkipSuperAdminLimit });
+const sessionCache = new Map();
+
+logService.attachConsole();
+
+registerCoreRoutes({ app, versionService });
+
+app.use(express.json({ limit: '1mb' }));
+
+registerPublicRoutes({
+  app,
+  dataStore,
+  mailerService,
+  logService,
+  escapeHtml,
+  sanitizeEmailLink
 });
 
-const loginLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 20,
-  message: { error: 'too_many_attempts', retryAfter: '15 minutes' },
-  standardHeaders: true,
-  legacyHeaders: false,
-  skipSuccessfulRequests: true,
-  keyGenerator: (req) => {
-    const teamName = typeof req.body?.teamName === 'string' ? req.body.teamName.toLowerCase() : '';
-    return `${ipKeyGenerator(req)}:${teamName}`;
-  }
+registerTeamRoutes({
+  app,
+  dataStore,
+  teamService,
+  tokenService,
+  mailerService,
+  logService,
+  escapeHtml,
+  authLimiter: rateLimiters.authLimiter,
+  loginLimiter: rateLimiters.loginLimiter,
+  teamReadLimiter: rateLimiters.teamReadLimiter,
+  teamWriteLimiter: rateLimiters.teamWriteLimiter
 });
 
-const superAdminActionLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 60,
-  message: { error: 'too_many_attempts', retryAfter: '15 minutes' },
-  standardHeaders: true,
-  legacyHeaders: false,
-  skip: shouldSkipSuperAdminLimit
+registerFeedbackRoutes({
+  app,
+  dataStore,
+  teamService,
+  mailerService,
+  logService,
+  escapeHtml,
+  teamReadLimiter: rateLimiters.teamReadLimiter,
+  teamWriteLimiter: rateLimiters.teamWriteLimiter
 });
 
-const teamReadLimiter = rateLimit({
-  windowMs: 60 * 1000,
-  max: 120,
-  message: { error: 'too_many_requests', retryAfter: '1 minute' },
-  standardHeaders: true,
-  legacyHeaders: false
+registerPasswordResetRoutes({
+  app,
+  dataStore,
+  mailerService,
+  escapeHtml,
+  sanitizeEmailLink,
+  hashResetToken,
+  pruneResetTokens
 });
 
-const teamWriteLimiter = rateLimit({
-  windowMs: 60 * 1000,
-  max: 60,
-  message: { error: 'too_many_requests', retryAfter: '1 minute' },
-  standardHeaders: true,
-  legacyHeaders: false
+registerSuperAdminRoutes({
+  app,
+  io,
+  dataStore,
+  tokenService,
+  mailerService,
+  logService,
+  escapeHtml,
+  superAdminPassword: SUPER_ADMIN_PASSWORD,
+  authLimiter: rateLimiters.authLimiter,
+  superAdminActionLimiter: rateLimiters.superAdminActionLimiter,
+  superAdminPollingLimiter: rateLimiters.superAdminPollingLimiter,
+  sessionCache
 });
 
-// ==================== DATABASE ABSTRACTION ====================
-// Supports PostgreSQL (multi-pod) and SQLite (single-pod/dev)
-// PostgreSQL config: DATABASE_URL or individual POSTGRES_* variables
-// If neither is set, SQLite is used
+registerSocketHandlers({ io, dataStore, sessionCache });
 
-const buildPostgresConfig = () => {
-  // Option 1: Full connection URL
-  if (process.env.DATABASE_URL) {
-    return { connectionString: process.env.DATABASE_URL };
-  }
-
-  // Option 2: Individual variables (for OpenShift secrets)
-  const host = process.env.POSTGRES_HOST || process.env.POSTGRESQL_SERVICE_HOST;
-  const port = process.env.POSTGRES_PORT || process.env.POSTGRESQL_SERVICE_PORT || 5432;
-  const user = process.env.POSTGRES_USER || process.env.POSTGRESQL_USER;
-  const password = process.env.POSTGRES_PASSWORD || process.env.POSTGRESQL_PASSWORD;
-  const database = process.env.POSTGRES_DB || process.env.POSTGRESQL_DATABASE;
-
-  if (host && user && password && database) {
-    return { host, port: Number(port), user, password, database };
-  }
-
-  return null;
-};
-
-const pgConfig = buildPostgresConfig();
-const usePostgres = !!pgConfig;
-
-// PostgreSQL setup
-let pgPool = null;
-
-const initPostgres = async () => {
-  const pool = new pg.Pool({
-    ...pgConfig,
-    max: 10,
-    idleTimeoutMillis: 30000,
-    connectionTimeoutMillis: 5000,
-  });
-  pool.on('error', (err) => {
-    console.error('[Server] Postgres pool error', err);
-  });
-
-  // Test connection
-  const client = await pool.connect();
-  try {
-    await client.query(`
-      CREATE TABLE IF NOT EXISTS kv_store (
-        key TEXT PRIMARY KEY,
-        value TEXT NOT NULL,
-        updated_at TIMESTAMPTZ NOT NULL DEFAULT NOW()
-      )
-    `);
-    console.info('[Server] Using PostgreSQL database (multi-pod ready)');
-  } finally {
-    client.release();
-  }
-
-  return pool;
-};
-
-// SQLite setup (fallback for single-pod/dev)
-const resolveDataStoreCandidates = () => {
-  const candidates = [];
-
-  if (process.env.DATA_STORE_PATH) {
-    candidates.push(process.env.DATA_STORE_PATH);
-  }
-
-  // Prefer a mounted volume when present (Railway/OpenShift)
-  candidates.push('/data/data.sqlite');
-
-  // As a last resort, write to /tmp (ephemeral but writable)
-  candidates.push(join('/tmp', 'data.sqlite'));
-
-  // Final fallback next to server.js (may be read-only in some images)
-  candidates.push(join(__dirname, 'data.sqlite'));
-
-  return candidates;
-};
-
-const openSqliteDatabase = () => {
-  const errors = [];
-
-  for (const candidate of resolveDataStoreCandidates()) {
-    try {
-      fs.mkdirSync(dirname(candidate), { recursive: true });
-      const database = new Database(candidate);
-      console.info(`[Server] Using SQLite store at ${candidate}`);
-
-      // Warn if using ephemeral storage
-      if (candidate.startsWith('/tmp')) {
-        console.warn('');
-        console.warn('┌──────────────────────────────────────────────────────────────────────┐');
-        console.warn('│ ⚠️  WARNING: Using ephemeral storage (/tmp)                           │');
-        console.warn('│    Data will be LOST when the container restarts!                    │');
-        console.warn('│                                                                      │');
-        console.warn('│    To persist data:                                                  │');
-        console.warn('│    - Railway: Add a Volume mounted at /data                          │');
-        console.warn('│    - Docker: Use -v /host/path:/data                                 │');
-        console.warn('│    - K8s/OpenShift: Create a PVC mounted at /data                    │');
-        console.warn('│    - Or set DATA_STORE_PATH to a persistent location                 │');
-        console.warn('└──────────────────────────────────────────────────────────────────────┘');
-        console.warn('');
-      }
-
-      return database;
-    } catch (err) {
-      errors.push({ pathTried: candidate, message: err?.message });
-      console.warn(`[Server] Failed to open SQLite store at ${candidate}: ${err?.message}`);
-    }
-  }
-
-  const error = new Error(
-    `Unable to open SQLite database. Paths tried: ${errors
-      .map((e) => `${e.pathTried} (${e.message})`)
-      .join('; ')}`
-  );
-  error.name = 'SQLiteInitError';
-  throw error;
-};
-
-let sqliteDb = null;
-
-const initSqlite = () => {
-  const db = openSqliteDatabase();
-  db.pragma('journal_mode = wal');
-  db.prepare(
-    `CREATE TABLE IF NOT EXISTS kv_store (
-      key TEXT PRIMARY KEY,
-      value TEXT NOT NULL,
-      updated_at TEXT NOT NULL DEFAULT CURRENT_TIMESTAMP
-    )`
-  ).run();
-  return db;
-};
-
-// Unified data access functions
-const normalizePersistedData = (data) => {
-  const normalized = data && typeof data === 'object' ? data : { teams: [] };
-  if (!Array.isArray(normalized.teams)) {
-    normalized.teams = [];
-  }
-
-  if (!normalized.meta || typeof normalized.meta !== 'object') {
-    normalized.meta = {
-      revision: 0,
-      updatedAt: new Date().toISOString()
-    };
-  } else {
-    if (typeof normalized.meta.revision !== 'number') {
-      normalized.meta.revision = 0;
-    }
-    if (!normalized.meta.updatedAt) {
-      normalized.meta.updatedAt = new Date().toISOString();
-    }
-  }
-
-  if (!Array.isArray(normalized.resetTokens)) {
-    normalized.resetTokens = [];
-  }
-
-  if (!Array.isArray(normalized.orphanedFeedbacks)) {
-    normalized.orphanedFeedbacks = [];
-  }
-
-  return normalized;
-};
-
-const bumpPersistedDataMeta = (data) => {
-  const normalized = normalizePersistedData(data);
-  return {
-    ...normalized,
-    meta: {
-      revision: (normalized.meta.revision || 0) + 1,
-      updatedAt: new Date().toISOString()
-    }
-  };
-};
-
-const loadPersistedData = async () => {
-  try {
-    if (usePostgres) {
-      const result = await pgPool.query('SELECT value FROM kv_store WHERE key = $1', ['retro-data']);
-      if (result.rows.length > 0 && result.rows[0].value) {
-        return normalizePersistedData(JSON.parse(result.rows[0].value));
-      }
-    } else {
-      const row = sqliteDb.prepare('SELECT value FROM kv_store WHERE key = ?').get('retro-data');
-      if (row?.value) {
-        return normalizePersistedData(JSON.parse(row.value));
-      }
-    }
-  } catch (err) {
-    console.warn('[Server] Failed to load persisted data store', err);
-  }
-  return normalizePersistedData({ teams: [] });
-};
-
-const savePersistedData = async (data) => {
-  const updatedData = bumpPersistedDataMeta(data);
-  const payload = JSON.stringify(updatedData ?? normalizePersistedData({ teams: [] }));
-
-  try {
-    if (usePostgres) {
-      await pgPool.query(
-        `INSERT INTO kv_store (key, value, updated_at)
-         VALUES ($1, $2, NOW())
-         ON CONFLICT (key) DO UPDATE SET
-           value = EXCLUDED.value,
-           updated_at = NOW()`,
-        ['retro-data', payload]
-      );
-    } else {
-      sqliteDb.prepare(
-        `INSERT INTO kv_store (key, value, updated_at)
-         VALUES (?, ?, CURRENT_TIMESTAMP)
-         ON CONFLICT(key) DO UPDATE SET
-           value = excluded.value,
-           updated_at = CURRENT_TIMESTAMP`
-      ).run('retro-data', payload);
-    }
-    return updatedData;
-  } catch (err) {
-    console.error('[Server] Failed to write persisted data store', err);
-    throw err;
-  }
-};
-
-/**
- * Atomically save data only if the expected revision matches.
- * Prevents TOCTOU race conditions across multiple pods.
- * Returns { success: true, data } on success, or { success: false, data } with
- * current server data on conflict.
- */
-const atomicSavePersistedData = async (incomingData, expectedRevision) => {
-  const nextData = bumpPersistedDataMeta({ ...incomingData, meta: { revision: expectedRevision, updatedAt: '' } });
-  const payload = JSON.stringify(nextData);
-
-  try {
-    if (usePostgres) {
-      const client = await pgPool.connect();
-      try {
-        await client.query('BEGIN');
-        const lockResult = await client.query(
-          'SELECT value FROM kv_store WHERE key = $1 FOR UPDATE',
-          ['retro-data']
-        );
-
-        const currentValue = lockResult.rows.length > 0 && lockResult.rows[0].value
-          ? normalizePersistedData(JSON.parse(lockResult.rows[0].value))
-          : normalizePersistedData({ teams: [] });
-        const serverRevision = Number(currentValue.meta?.revision ?? 0);
-
-        if (expectedRevision !== serverRevision) {
-          await client.query('ROLLBACK');
-          return { success: false, data: currentValue };
-        }
-
-        await client.query(
-          `INSERT INTO kv_store (key, value, updated_at)
-           VALUES ($1, $2, NOW())
-           ON CONFLICT (key) DO UPDATE SET
-             value = EXCLUDED.value,
-             updated_at = NOW()`,
-          ['retro-data', payload]
-        );
-        await client.query('COMMIT');
-        return { success: true, data: nextData };
-      } catch (txErr) {
-        await client.query('ROLLBACK').catch(() => {});
-        throw txErr;
-      } finally {
-        client.release();
-      }
-    } else {
-      // SQLite: use a synchronous transaction (better-sqlite3 transactions are atomic)
-      const result = sqliteDb.transaction(() => {
-        const row = sqliteDb.prepare('SELECT value FROM kv_store WHERE key = ?').get('retro-data');
-        const currentValue = row?.value
-          ? normalizePersistedData(JSON.parse(row.value))
-          : normalizePersistedData({ teams: [] });
-        const serverRevision = Number(currentValue.meta?.revision ?? 0);
-
-        if (expectedRevision !== serverRevision) {
-          return { success: false, data: currentValue };
-        }
-
-        sqliteDb.prepare(
-          `INSERT INTO kv_store (key, value, updated_at)
-           VALUES (?, ?, CURRENT_TIMESTAMP)
-           ON CONFLICT(key) DO UPDATE SET
-             value = excluded.value,
-             updated_at = CURRENT_TIMESTAMP`
-        ).run('retro-data', payload);
-        return { success: true, data: nextData };
-      })();
-      return result;
-    }
-  } catch (err) {
-    console.error('[Server] Failed atomic save', err);
-    throw err;
-  }
-};
-
-const loadSessionState = async (sessionId) => {
-  const key = `session:${sessionId}`;
-
-  try {
-    if (usePostgres) {
-      const result = await pgPool.query('SELECT value FROM kv_store WHERE key = $1', [key]);
-      if (result.rows.length > 0 && result.rows[0].value) {
-        return JSON.parse(result.rows[0].value);
-      }
-    } else {
-      const row = sqliteDb.prepare('SELECT value FROM kv_store WHERE key = ?').get(key);
-      if (row?.value) {
-        return JSON.parse(row.value);
-      }
-    }
-  } catch (err) {
-    console.warn('[Server] Failed to load session state', err);
-  }
-
-  return null;
-};
-
-const saveSessionState = async (sessionId, sessionData) => {
-  const key = `session:${sessionId}`;
-
-  // Bump session revision for conflict tracking
-  const currentRev = Number(sessionData?._rev ?? 0);
-  const dataWithRev = { ...sessionData, _rev: currentRev + 1, _updatedAt: new Date().toISOString() };
-  const payload = JSON.stringify(dataWithRev);
-
-  try {
-    if (usePostgres) {
-      await pgPool.query(
-        `INSERT INTO kv_store (key, value, updated_at)
-         VALUES ($1, $2, NOW())
-         ON CONFLICT (key) DO UPDATE SET
-           value = EXCLUDED.value,
-           updated_at = NOW()`,
-        [key, payload]
-      );
-    } else {
-      sqliteDb.prepare(
-        `INSERT INTO kv_store (key, value, updated_at)
-         VALUES (?, ?, CURRENT_TIMESTAMP)
-         ON CONFLICT(key) DO UPDATE SET
-           value = excluded.value,
-           updated_at = CURRENT_TIMESTAMP`
-      ).run(key, payload);
-    }
-    return dataWithRev;
-  } catch (err) {
-    console.error('[Server] Failed to write session state', err);
-    throw err;
-  }
-};
-
-const refreshPersistedData = async () => {
-  persistedData = await loadPersistedData();
-  return persistedData;
-};
-
-// Global settings functions
-const loadGlobalSettings = async () => {
-  try {
-    if (usePostgres) {
-      const result = await pgPool.query('SELECT value FROM kv_store WHERE key = $1', ['global-settings']);
-      if (result.rows.length > 0 && result.rows[0].value) {
-        return JSON.parse(result.rows[0].value);
-      }
-    } else {
-      const row = sqliteDb.prepare('SELECT value FROM kv_store WHERE key = ?').get('global-settings');
-      if (row?.value) {
-        return JSON.parse(row.value);
-      }
-    }
-  } catch (err) {
-    console.warn('[Server] Failed to load global settings', err);
-  }
-  return {};
-};
-
-const saveGlobalSettings = async (settings) => {
-  const payload = JSON.stringify(settings ?? {});
-
-  try {
-    if (usePostgres) {
-      await pgPool.query(
-        `INSERT INTO kv_store (key, value, updated_at)
-         VALUES ($1, $2, NOW())
-         ON CONFLICT (key) DO UPDATE SET
-           value = EXCLUDED.value,
-           updated_at = NOW()`,
-        ['global-settings', payload]
-      );
-    } else {
-      sqliteDb.prepare(
-        `INSERT INTO kv_store (key, value, updated_at)
-         VALUES (?, ?, CURRENT_TIMESTAMP)
-         ON CONFLICT(key) DO UPDATE SET
-           value = excluded.value,
-           updated_at = CURRENT_TIMESTAMP`
-      ).run('global-settings', payload);
-    }
-  } catch (err) {
-    console.error('[Server] Failed to write global settings', err);
-    throw err;
-  }
-};
-
-// Initialize database based on configuration
-const initDatabase = async () => {
-  if (usePostgres) {
-    pgPool = await initPostgres();
-  } else {
-    sqliteDb = initSqlite();
-  }
-};
-
-// Will be populated after database init
-let persistedData = normalizePersistedData({ teams: [] });
-
-// ==================== SERVER LOGS STORAGE ====================
-// In-memory circular buffer for server logs (keeps last 500 entries)
-const MAX_LOG_ENTRIES = 500;
-const serverLogs = [];
-let logIdCounter = 0;
-
-const addServerLog = (level, source, message, details = null) => {
-  const entry = {
-    id: String(++logIdCounter),
-    timestamp: new Date().toISOString(),
-    level,
-    source,
-    message,
-    details: details || undefined
-  };
-  serverLogs.push(entry);
-  // Keep only last MAX_LOG_ENTRIES
-  if (serverLogs.length > MAX_LOG_ENTRIES) {
-    serverLogs.shift();
-  }
-  return entry;
-};
-
-// Override console methods to capture logs
-const originalConsoleError = console.error;
-const originalConsoleWarn = console.warn;
-
-console.error = (...args) => {
-  originalConsoleError.apply(console, args);
-  const message = args.map(a => typeof a === 'object' ? JSON.stringify(a) : String(a)).join(' ');
-  // Determine source from message
-  let source = 'server';
-  if (message.includes('[Postgres]') || message.includes('postgres') || message.includes('pg_')) {
-    source = 'postgres';
-  } else if (message.includes('[Socket') || message.includes('Socket IO')) {
-    source = 'socket';
-  } else if (message.includes('email') || message.includes('SMTP') || message.includes('mailer')) {
-    source = 'email';
-  }
-  addServerLog('error', source, message.substring(0, 500));
-};
-
-console.warn = (...args) => {
-  originalConsoleWarn.apply(console, args);
-  const message = args.map(a => typeof a === 'object' ? JSON.stringify(a) : String(a)).join(' ');
-  let source = 'server';
-  if (message.includes('[Postgres]') || message.includes('postgres')) {
-    source = 'postgres';
-  } else if (message.includes('[Socket') || message.includes('Socket IO')) {
-    source = 'socket';
-  }
-  addServerLog('warn', source, message.substring(0, 500));
-};
-
-const smtpEnabled = !!process.env.SMTP_HOST;
-const mailer = smtpEnabled
-  ? nodemailer.createTransport({
-      host: process.env.SMTP_HOST,
-      port: Number(process.env.SMTP_PORT || 587),
-      secure: process.env.SMTP_SECURE === 'true',
-      auth: process.env.SMTP_USER
-        ? { user: process.env.SMTP_USER, pass: process.env.SMTP_PASS }
-        : undefined
-    })
-  : null;
-
-const RESET_TOKEN_TTL_MS = 60 * 60 * 1000;
-
-const hashResetToken = (token) =>
-  createHash('sha256').update(token).digest('hex');
-
-const pruneResetTokens = (tokens) => {
-  const now = Date.now();
-  return (tokens || []).filter((entry) => entry.expiresAt > now);
-};
-
-// ==================== SECURE TEAM-SCOPED API ====================
-// These endpoints replace the old /api/data which exposed all team data
-
-/**
- * Sanitize a team object for client response - NEVER expose passwords
- * or data from other teams
- */
-const sanitizeTeamForClient = (team) => {
-  if (!team) return null;
-  const { passwordHash, ...safeTeam } = team;
-  return safeTeam;
-};
-
-/**
- * Find a team by ID and verify password
- * Returns { team, error } where team is the full team object if auth succeeds
- */
-const authenticateTeam = async (teamId, password) => {
-  const currentData = await loadPersistedData();
-  const team = currentData.teams.find(t => t.id === teamId);
-
-  if (!team) {
-    return { team: null, error: 'team_not_found' };
-  }
-
-  if (!password || team.passwordHash !== password) {
-    return { team: null, error: 'invalid_password' };
-  }
-
-  return { team, error: null, currentData };
-};
-
-/**
- * Atomically update a single team
- */
-const atomicUpdateTeam = async (teamId, updater) => {
-  const MAX_RETRIES = 5;
-  for (let attempt = 0; attempt < MAX_RETRIES; attempt++) {
-    const currentData = await loadPersistedData();
-    const teamIndex = currentData.teams.findIndex(t => t.id === teamId);
-
-    if (teamIndex === -1) {
-      return { success: false, error: 'team_not_found' };
-    }
-
-    const updatedTeam = updater(currentData.teams[teamIndex]);
-    if (!updatedTeam) {
-      return { success: true, team: currentData.teams[teamIndex] };
-    }
-
-    currentData.teams[teamIndex] = updatedTeam;
-    const revision = Number(currentData.meta?.revision ?? 0);
-    const result = await atomicSavePersistedData(currentData, revision);
-
-    if (result.success) {
-      persistedData = result.data;
-      return { success: true, team: updatedTeam };
-    }
-
-    console.warn(`[Server] Team update conflict, retry ${attempt + 1}/${MAX_RETRIES}`);
-  }
-
-  return { success: false, error: 'max_retries_exceeded' };
-};
-
-/**
- * Helper for operations: atomically read-modify-write with retry.
- * The mutator function receives the current data and must return the modified copy.
- * Returns the updated data or throws on failure.
- */
-const atomicReadModifyWrite = async (mutator) => {
-  const MAX_RETRIES = 5;
-  for (let attempt = 0; attempt < MAX_RETRIES; attempt++) {
-    const currentData = await loadPersistedData();
-    const modified = mutator(currentData);
-    if (!modified) return currentData; // mutator can return null to signal no-op
-
-    const revision = Number(currentData.meta?.revision ?? 0);
-    const result = await atomicSavePersistedData(modified, revision);
-    if (result.success) {
-      persistedData = result.data;
-      return result.data;
-    }
-    // Conflict - retry with fresh data
-    console.warn(`[Server] Operation conflict, retry ${attempt + 1}/${MAX_RETRIES}`);
-  }
-  throw new Error('Failed to save after max retries');
-};
-
-// POST /api/team/login - Authenticate and get team data
-app.post('/api/team/login', loginLimiter, async (req, res) => {
-  try {
-    const { teamName, password } = req.body || {};
-
-    if (!teamName || !password) {
-      return res.status(400).json({ error: 'missing_credentials' });
-    }
-
-    const currentData = await loadPersistedData();
-    const team = currentData.teams.find(t =>
-      t.name.toLowerCase() === teamName.toLowerCase()
-    );
-
-    if (!team) {
-      return res.status(401).json({ error: 'team_not_found' });
-    }
-
-    if (team.passwordHash !== password) {
-      return res.status(401).json({ error: 'invalid_password' });
-    }
-
-    // Generate session token for secure browser refresh persistence
-    const sessionToken = createSessionToken(team.id, null);
-
-    res.json({
-      team: sanitizeTeamForClient(team),
-      meta: currentData.meta,
-      sessionToken
-    });
-  } catch (err) {
-    console.error('[Server] Failed to login team', err);
-    res.status(500).json({ error: 'login_failed' });
-  }
-});
-
-// POST /api/team/restore-session - Restore session using a token (no password needed)
-app.post('/api/team/restore-session', authLimiter, async (req, res) => {
-  try {
-    const { sessionToken } = req.body || {};
-
-    if (!sessionToken) {
-      return res.status(400).json({ error: 'missing_token' });
-    }
-
-    const session = validateSessionToken(sessionToken);
-    if (!session) {
-      return res.status(401).json({ error: 'invalid_or_expired_token' });
-    }
-
-    const currentData = await loadPersistedData();
-    const team = currentData.teams.find(t => t.id === session.teamId);
-
-    if (!team) {
-      invalidateSessionToken(sessionToken);
-      return res.status(404).json({ error: 'team_not_found' });
-    }
-
-    res.json({
-      team: sanitizeTeamForClient(team),
-      meta: currentData.meta,
-      password: team.passwordHash // Needed for creating invite links
-    });
-  } catch (err) {
-    console.error('[Server] Failed to restore session', err);
-    res.status(500).json({ error: 'restore_failed' });
-  }
-});
-
-// POST /api/team/create - Create a new team
-app.post('/api/team/create', authLimiter, async (req, res) => {
-  try {
-    const { name, password, facilitatorEmail } = req.body || {};
-
-    if (!name || !password) {
-      return res.status(400).json({ error: 'missing_fields' });
-    }
-
-    if (password.length < 4) {
-      return res.status(400).json({ error: 'password_too_short' });
-    }
-
-    const MAX_RETRIES = 5;
-    for (let attempt = 0; attempt < MAX_RETRIES; attempt++) {
-      const currentData = await loadPersistedData();
-
-      // Check if team name already exists
-      if (currentData.teams.some(t => t.name.toLowerCase() === name.toLowerCase())) {
-        return res.status(409).json({ error: 'team_name_exists' });
-      }
-
-      const newTeam = {
-        id: Math.random().toString(36).substr(2, 9),
-        name,
-        passwordHash: password,
-        facilitatorEmail: facilitatorEmail || undefined,
-        members: [
-          {
-            id: 'admin-' + Math.random().toString(36).substr(2, 5),
-            name: 'Facilitator',
-            color: 'bg-indigo-500',
-            role: 'facilitator'
-          }
-        ],
-        archivedMembers: [],
-        customTemplates: [],
-        retrospectives: [],
-        globalActions: []
-      };
-
-      currentData.teams.push(newTeam);
-      const revision = Number(currentData.meta?.revision ?? 0);
-      const result = await atomicSavePersistedData(currentData, revision);
-
-      if (result.success) {
-        persistedData = result.data;
-
-        // Send new team notification email if enabled
-        if (smtpEnabled && mailer) {
-          try {
-            const settings = await loadGlobalSettings();
-            if (settings.notifyNewTeam && settings.adminEmail) {
-              const safeTeamName = escapeHtml(newTeam.name);
-              const createdAt = new Date().toLocaleString();
-              await mailer.sendMail({
-                from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-                to: settings.adminEmail,
-                subject: `New team created: ${newTeam.name}`,
-                html: `
-                  <div style="font-family: Arial, sans-serif; max-width: 500px; margin: 0 auto; padding: 20px;">
-                    <h2 style="color: #334155;">New Team Created</h2>
-                    <div style="background: #f1f5f9; border-radius: 8px; padding: 16px; margin: 16px 0;">
-                      <p style="margin: 0 0 8px 0;"><strong>Team name:</strong> ${safeTeamName}</p>
-                      <p style="margin: 0 0 8px 0;"><strong>Team ID:</strong> ${newTeam.id}</p>
-                      <p style="margin: 0;"><strong>Created at:</strong> ${createdAt}</p>
-                    </div>
-                    <p style="color: #64748b; font-size: 12px;">This is an automated notification from RetroGemini.</p>
-                  </div>
-                `,
-                text: `New team created:\n\nTeam name: ${newTeam.name}\nTeam ID: ${newTeam.id}\nCreated at: ${createdAt}`
-              });
-              addServerLog('info', 'email', `New team notification sent to ${settings.adminEmail} for team: ${newTeam.name}`);
-            }
-          } catch (emailErr) {
-            addServerLog('warn', 'email', `Failed to send new team notification: ${emailErr.message}`);
-          }
-        }
-
-        return res.status(201).json({
-          team: sanitizeTeamForClient(newTeam),
-          meta: result.data.meta
-        });
-      }
-
-      console.warn(`[Server] Team create conflict, retry ${attempt + 1}/${MAX_RETRIES}`);
-    }
-
-    res.status(500).json({ error: 'failed_to_create' });
-  } catch (err) {
-    console.error('[Server] Failed to create team', err);
-    res.status(500).json({ error: 'failed_to_create' });
-  }
-});
-
-// GET /api/team/list - List team names for login (no sensitive data)
-app.get('/api/team/list', teamReadLimiter, async (_req, res) => {
-  try {
-    const currentData = await loadPersistedData();
-    const teams = currentData.teams.map((team) => ({
-      id: team.id,
-      name: team.name,
-      memberCount: Array.isArray(team.members) ? team.members.length : 0,
-      lastConnectionDate: team.lastConnectionDate
-    })).sort((a, b) => a.name.localeCompare(b.name, undefined, { sensitivity: 'base' }));
-    res.json({ teams });
-  } catch (err) {
-    console.error('[Server] Failed to list teams', err);
-    res.status(500).json({ error: 'failed_to_list' });
-  }
-});
-
-// POST /api/team/:teamId - Get team data (with auth in body for security)
-app.post('/api/team/:teamId', teamReadLimiter, async (req, res) => {
-  try {
-    const { teamId } = req.params;
-    const { password } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    const currentData = await loadPersistedData();
-    res.json({
-      team: sanitizeTeamForClient(team),
-      meta: currentData.meta
-    });
-  } catch (err) {
-    console.error('[Server] Failed to get team', err);
-    res.status(500).json({ error: 'failed_to_load' });
-  }
-});
-
-// POST /api/team/:teamId/update - Update team data (partial update)
-app.post('/api/team/:teamId/update', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId } = req.params;
-    const { password, updates } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!updates || typeof updates !== 'object') {
-      return res.status(400).json({ error: 'invalid_updates' });
-    }
-
-    const result = await atomicUpdateTeam(teamId, (currentTeam) => {
-      // Merge updates into the team, but NEVER allow updating the password via this endpoint
-      const { passwordHash, id, ...safeUpdates } = updates;
-      return {
-        ...currentTeam,
-        ...safeUpdates,
-        id: currentTeam.id, // Prevent ID change
-        passwordHash: currentTeam.passwordHash // Prevent password change via this endpoint
-      };
-    });
-
-    if (!result.success) {
-      return res.status(500).json({ error: result.error });
-    }
-
-    const currentData = await loadPersistedData();
-    res.json({
-      team: sanitizeTeamForClient(result.team),
-      meta: currentData.meta
-    });
-  } catch (err) {
-    console.error('[Server] Failed to update team', err);
-    res.status(500).json({ error: 'failed_to_update' });
-  }
-});
-
-// POST /api/team/:teamId/retrospective/:retroId - Update a specific retrospective
-app.post('/api/team/:teamId/retrospective/:retroId', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId, retroId } = req.params;
-    const { password, retrospective } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!retrospective) {
-      return res.status(400).json({ error: 'missing_retrospective' });
-    }
-
-    const result = await atomicUpdateTeam(teamId, (currentTeam) => {
-      if (!currentTeam.retrospectives) currentTeam.retrospectives = [];
-
-      const idx = currentTeam.retrospectives.findIndex(r => r.id === retroId);
-      if (idx !== -1) {
-        currentTeam.retrospectives[idx] = { ...retrospective, id: retroId, teamId };
-      } else {
-        currentTeam.retrospectives.unshift({ ...retrospective, id: retroId, teamId });
-      }
-
-      return currentTeam;
-    });
-
-    if (!result.success) {
-      return res.status(500).json({ error: result.error });
-    }
-
-    const currentData = await loadPersistedData();
-    res.json({ meta: currentData.meta });
-  } catch (err) {
-    console.error('[Server] Failed to update retrospective', err);
-    res.status(500).json({ error: 'failed_to_update' });
-  }
-});
-
-// POST /api/team/:teamId/healthcheck/:hcId - Update a specific health check
-app.post('/api/team/:teamId/healthcheck/:hcId', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId, hcId } = req.params;
-    const { password, healthCheck } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!healthCheck) {
-      return res.status(400).json({ error: 'missing_healthcheck' });
-    }
-
-    const result = await atomicUpdateTeam(teamId, (currentTeam) => {
-      if (!currentTeam.healthChecks) currentTeam.healthChecks = [];
-
-      const idx = currentTeam.healthChecks.findIndex(h => h.id === hcId);
-      if (idx !== -1) {
-        currentTeam.healthChecks[idx] = { ...healthCheck, id: hcId, teamId };
-      } else {
-        currentTeam.healthChecks.unshift({ ...healthCheck, id: hcId, teamId });
-      }
-
-      return currentTeam;
-    });
-
-    if (!result.success) {
-      return res.status(500).json({ error: result.error });
-    }
-
-    const currentData = await loadPersistedData();
-    res.json({ meta: currentData.meta });
-  } catch (err) {
-    console.error('[Server] Failed to update health check', err);
-    res.status(500).json({ error: 'failed_to_update' });
-  }
-});
-
-// POST /api/team/:teamId/action - Update or create an action
-app.post('/api/team/:teamId/action', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId } = req.params;
-    const { password, action, retroId } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!action || !action.id) {
-      return res.status(400).json({ error: 'missing_action' });
-    }
-
-    const result = await atomicUpdateTeam(teamId, (currentTeam) => {
-      if (!currentTeam.globalActions) currentTeam.globalActions = [];
-
-      // Try to update in global actions first
-      const globalIdx = currentTeam.globalActions.findIndex(a => a.id === action.id);
-      if (globalIdx !== -1) {
-        currentTeam.globalActions[globalIdx] = { ...action };
-        return currentTeam;
-      }
-
-      // Try to update in retrospective actions
-      if (retroId && currentTeam.retrospectives) {
-        const retro = currentTeam.retrospectives.find(r => r.id === retroId);
-        if (retro && retro.actions) {
-          const retroActionIdx = retro.actions.findIndex(a => a.id === action.id);
-          if (retroActionIdx !== -1) {
-            retro.actions[retroActionIdx] = { ...action };
-            return currentTeam;
-          }
-        }
-      }
-
-      // If not found anywhere, add to global actions
-      currentTeam.globalActions.unshift(action);
-      return currentTeam;
-    });
-
-    if (!result.success) {
-      return res.status(500).json({ error: result.error });
-    }
-
-    const currentData = await loadPersistedData();
-    res.json({ meta: currentData.meta });
-  } catch (err) {
-    console.error('[Server] Failed to update action', err);
-    res.status(500).json({ error: 'failed_to_update' });
-  }
-});
-
-// POST /api/team/:teamId/members - Update team members
-app.post('/api/team/:teamId/members', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId } = req.params;
-    const { password, members, archivedMembers } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    const result = await atomicUpdateTeam(teamId, (currentTeam) => {
-      if (members !== undefined) {
-        currentTeam.members = members;
-      }
-      if (archivedMembers !== undefined) {
-        currentTeam.archivedMembers = archivedMembers;
-      }
-      return currentTeam;
-    });
-
-    if (!result.success) {
-      return res.status(500).json({ error: result.error });
-    }
-
-    const currentData = await loadPersistedData();
-    res.json({
-      team: sanitizeTeamForClient(result.team),
-      meta: currentData.meta
-    });
-  } catch (err) {
-    console.error('[Server] Failed to update members', err);
-    res.status(500).json({ error: 'failed_to_update' });
-  }
-});
-
-// POST /api/team/:teamId/password - Change team password
-app.post('/api/team/:teamId/password', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId } = req.params;
-    const { password, newPassword } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!newPassword || newPassword.length < 4) {
-      return res.status(400).json({ error: 'password_too_short' });
-    }
-
-    const result = await atomicUpdateTeam(teamId, (currentTeam) => {
-      currentTeam.passwordHash = newPassword;
-      return currentTeam;
-    });
-
-    if (!result.success) {
-      return res.status(500).json({ error: result.error });
-    }
-
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to change password', err);
-    res.status(500).json({ error: 'failed_to_update' });
-  }
-});
-
-// POST /api/team/:teamId/delete - Delete a team
-app.post('/api/team/:teamId/delete', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId } = req.params;
-    const { password } = req.body || {};
-
-    const { team, error } = await authenticateTeam(teamId, password);
-
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    const MAX_RETRIES = 5;
-    for (let attempt = 0; attempt < MAX_RETRIES; attempt++) {
-      const currentData = await loadPersistedData();
-      const idx = currentData.teams.findIndex(t => t.id === teamId);
-
-      if (idx === -1) {
-        return res.status(404).json({ error: 'team_not_found' });
-      }
-
-      // Preserve feedbacks from the deleted team
-      const deletedTeam = currentData.teams[idx];
-      if (deletedTeam.teamFeedbacks && deletedTeam.teamFeedbacks.length > 0) {
-        if (!Array.isArray(currentData.orphanedFeedbacks)) {
-          currentData.orphanedFeedbacks = [];
-        }
-        const feedbacksToPreserve = deletedTeam.teamFeedbacks.map(f => ({
-          ...f,
-          teamId: f.teamId || deletedTeam.id,
-          teamName: f.teamName || deletedTeam.name
-        }));
-        currentData.orphanedFeedbacks.push(...feedbacksToPreserve);
-      }
-
-      currentData.teams.splice(idx, 1);
-      const revision = Number(currentData.meta?.revision ?? 0);
-      const result = await atomicSavePersistedData(currentData, revision);
-
-      if (result.success) {
-        persistedData = result.data;
-        return res.json({ success: true });
-      }
-
-      console.warn(`[Server] Team delete conflict, retry ${attempt + 1}/${MAX_RETRIES}`);
-    }
-
-    res.status(500).json({ error: 'failed_to_delete' });
-  } catch (err) {
-    console.error('[Server] Failed to delete team', err);
-    res.status(500).json({ error: 'failed_to_delete' });
-  }
-});
-
-// GET /api/team/exists/:teamName - Check if a team exists (for team creation UX)
-app.get('/api/team/exists/:teamName', async (req, res) => {
-  try {
-    const { teamName } = req.params;
-    const currentData = await loadPersistedData();
-    const exists = currentData.teams.some(t =>
-      t.name.toLowerCase() === decodeURIComponent(teamName).toLowerCase()
-    );
-    res.json({ exists });
-  } catch (err) {
-    console.error('[Server] Failed to check team existence', err);
-    res.status(500).json({ error: 'check_failed' });
-  }
-});
-
-// DEPRECATED: Old /api/data endpoints - Remove in future version
-// Kept temporarily for backwards compatibility during rollout
-app.get('/api/data', async (_req, res) => {
-  // Return empty data - forces clients to use new secure endpoints
-  console.warn('[Server] DEPRECATED: /api/data GET called - client should use /api/team endpoints');
-  res.status(410).json({ error: 'endpoint_deprecated', teams: [], meta: { revision: 0, updatedAt: new Date().toISOString() } });
-});
-
-app.post('/api/data', async (_req, res) => {
-  // Reject all writes through old endpoint
-  console.warn('[Server] DEPRECATED: /api/data POST called - client should use /api/team endpoints');
-  res.status(410).json({ error: 'endpoint_deprecated', message: 'Use /api/team endpoints instead' });
-});
-
-app.post('/api/send-invite', async (req, res) => {
-  if (!smtpEnabled || !mailer) {
-    return res.status(501).json({ error: 'email_not_configured' });
-  }
-
-  const { email, name, link, teamName, sessionName } = req.body || {};
-  if (!email || !link) {
-    return res.status(400).json({ error: 'missing_fields' });
-  }
-
-  const compactedLink = compactInviteLink(link);
-  const safeInviteLink = sanitizeEmailLink(compactedLink);
-  const safeName = escapeHtml(name || 'You');
-  const safeTeamName = escapeHtml(teamName || 'a RetroGemini team');
-  const safeSessionName = sessionName ? escapeHtml(sessionName) : '';
-  const safeInviteLinkHtml = escapeHtml(safeInviteLink);
-
-  try {
-    await mailer.sendMail({
-      from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-      to: email,
-      subject: `Invitation to join ${teamName || 'RetroGemini'}`,
-      text: `${name || 'You'},
-
-You have been invited to join ${teamName || 'a RetroGemini team'}${sessionName ? ` for the session "${sessionName}"` : ''}.
-Use this link to join: ${compactedLink}
-`,
-      html: `<p>${safeName},</p>
-<p>You have been invited to join <strong>${safeTeamName}</strong>${safeSessionName ? ` for the session "${safeSessionName}"` : ''}.</p>
-<p><a href="${safeInviteLinkHtml}" target="_blank" rel="noreferrer">Join with this link</a></p>`
-    });
-
-    res.status(204).end();
-  } catch (err) {
-    console.error('[Server] Failed to send invite email', err);
-    res.status(500).json({ error: 'send_failed' });
-  }
-});
-
-// Public endpoint to send feedback notification email to admin (called when feedback is submitted)
-app.post('/api/notify-new-feedback', async (req, res) => {
-  // This endpoint doesn't require authentication - it just sends a notification
-  // if admin email is configured
-
-  if (!smtpEnabled || !mailer) {
-    // Silently succeed if email not configured - this is expected in many deployments
-    return res.status(204).end();
-  }
-
-  const { feedback } = req.body || {};
-
-  try {
-    const settings = await loadGlobalSettings();
-    const adminEmail = settings.adminEmail;
-
-    if (!adminEmail) {
-      // No admin email configured - silently succeed
-      return res.status(204).end();
-    }
-
-    if (!feedback || !feedback.title || !feedback.type) {
-      return res.status(400).json({ error: 'missing_feedback_data' });
-    }
-
-    const typeLabel = feedback.type === 'bug' ? 'Bug Report' : 'Feature Request';
-    const typeEmoji = feedback.type === 'bug' ? '🐛' : '✨';
-    const safeFeedbackTitle = escapeHtml(feedback.title);
-    const safeFeedbackTeamName = escapeHtml(feedback.teamName);
-    const safeFeedbackSubmittedBy = escapeHtml(feedback.submittedByName);
-    const safeFeedbackDescription = escapeHtml(feedback.description);
-    const feedbackDate = new Date(feedback.submittedAt).toLocaleString();
-
-    await mailer.sendMail({
-      from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-      to: adminEmail,
-      subject: `${typeEmoji} New ${typeLabel}: ${feedback.title}`,
-      text: `New ${typeLabel} submitted
-
-Title: ${feedback.title}
-Type: ${typeLabel}
-Team: ${feedback.teamName}
-Submitted by: ${feedback.submittedByName}
-Date: ${feedbackDate}
-
-Description:
-${feedback.description}
-
----
-Log in to the Super Admin Dashboard to review and respond to this feedback.
-`,
-      html: `
-<div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-  <h2 style="color: ${feedback.type === 'bug' ? '#dc2626' : '#7c3aed'};">
-    ${typeEmoji} New ${typeLabel}
-  </h2>
-  <div style="background: #f8fafc; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <h3 style="margin: 0 0 8px 0; color: #1e293b;">${safeFeedbackTitle}</h3>
-    <p style="margin: 4px 0; color: #64748b; font-size: 14px;">
-      <strong>Team:</strong> ${safeFeedbackTeamName}<br>
-      <strong>Submitted by:</strong> ${safeFeedbackSubmittedBy}<br>
-      <strong>Date:</strong> ${feedbackDate}
-    </p>
-  </div>
-  <div style="margin: 16px 0;">
-    <h4 style="color: #475569; margin-bottom: 8px;">Description:</h4>
-    <p style="color: #334155; white-space: pre-wrap;">${safeFeedbackDescription}</p>
-  </div>
-  ${feedback.images && feedback.images.length > 0 ? `
-  <p style="color: #64748b; font-size: 14px;">
-    <em>${feedback.images.length} image(s) attached - view in Super Admin Dashboard</em>
-  </p>
-  ` : ''}
-  <hr style="border: none; border-top: 1px solid #e2e8f0; margin: 24px 0;">
-  <p style="color: #94a3b8; font-size: 12px;">
-    Log in to the Super Admin Dashboard to review and respond to this feedback.
-  </p>
-</div>
-`
-    });
-
-    addServerLog('info', 'email', `Feedback notification sent to ${adminEmail} for: ${feedback.title}`);
-    res.status(204).end();
-  } catch (err) {
-    console.error('[Server] Failed to send feedback notification email', err);
-    // Don't fail the request if email fails - feedback was already saved
-    res.status(204).end();
-  }
-});
-
-// Create a new feedback (server-side to avoid sync issues)
-app.post('/api/feedbacks/create', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId, password, feedback } = req.body || {};
-
-    // Authenticate the requesting team
-    const { team, error } = await authenticateTeam(teamId, password);
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!feedback || !feedback.type || !feedback.title || !feedback.description) {
-      return res.status(400).json({ error: 'missing_feedback_data' });
-    }
-
-    const feedbackId = 'feedback_' + Math.random().toString(36).substr(2, 9);
-    const newFeedback = {
-      id: feedbackId,
-      teamId,
-      teamName: team.name,
-      type: feedback.type,
-      title: feedback.title.trim().slice(0, 100),
-      description: feedback.description.trim().slice(0, 2000),
-      images: feedback.images || undefined,
-      submittedBy: feedback.submittedBy,
-      submittedByName: feedback.submittedByName,
-      submittedAt: new Date().toISOString(),
-      isRead: false,
-      status: 'pending',
-      comments: []
-    };
-
-    await atomicReadModifyWrite((data) => {
-      const targetTeam = data.teams.find(t => t.id === teamId);
-      if (!targetTeam) return null;
-      if (!targetTeam.teamFeedbacks) {
-        targetTeam.teamFeedbacks = [];
-      }
-      targetTeam.teamFeedbacks.unshift(newFeedback);
-      return data;
-    });
-
-    res.json({ success: true, feedback: newFeedback });
-  } catch (err) {
-    console.error('[Server] Failed to create feedback', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-// Get all feedbacks from all teams (for shared feedback hub)
-app.post('/api/feedbacks/all', teamReadLimiter, async (req, res) => {
-  try {
-    const { teamId, password } = req.body || {};
-
-    // Authenticate the requesting team
-    const { error } = await authenticateTeam(teamId, password);
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    const currentData = await loadPersistedData();
-    const feedbacks = currentData.teams.flatMap((team) =>
-      (team.teamFeedbacks || []).map((feedback) => ({
-        ...feedback,
-        teamId: feedback.teamId || team.id,
-        teamName: feedback.teamName || team.name,
-        isRead: feedback.isRead ?? false,
-        status: feedback.status || 'pending',
-        comments: feedback.comments || []
-      }))
-    );
-    // Include orphaned feedbacks from deleted teams
-    const orphaned = (currentData.orphanedFeedbacks || []).map((feedback) => ({
-      ...feedback,
-      isRead: feedback.isRead ?? false,
-      status: feedback.status || 'pending',
-      comments: feedback.comments || []
-    }));
-    feedbacks.push(...orphaned);
-    feedbacks.sort((a, b) => new Date(b.submittedAt).getTime() - new Date(a.submittedAt).getTime());
-    res.json({ feedbacks });
-  } catch (err) {
-    console.error('[Server] Failed to load all feedbacks', err);
-    res.status(500).json({ error: 'failed_to_load' });
-  }
-});
-
-// Add a comment to a feedback
-app.post('/api/feedbacks/comment', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId, password, feedbackTeamId, feedbackId, authorId, authorName, content } = req.body || {};
-
-    // Authenticate the requesting team
-    const { error } = await authenticateTeam(teamId, password);
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!feedbackTeamId || !feedbackId || !authorId || !authorName || !content) {
-      return res.status(400).json({ error: 'missing_comment_data' });
-    }
-
-    const currentData = await loadPersistedData();
-    const requestingTeam = currentData.teams.find(t => t.id === teamId);
-    const requestingTeamName = requestingTeam ? requestingTeam.name : 'Unknown Team';
-
-    const commentId = `comment_${Date.now()}_${Math.random().toString(36).substr(2, 5)}`;
-    const newComment = {
-      id: commentId,
-      feedbackId,
-      teamId,
-      teamName: requestingTeamName,
-      authorId,
-      authorName,
-      content: content.trim().slice(0, 1000), // Limit to 1000 chars
-      createdAt: new Date().toISOString()
-    };
-
-    let feedbackTitle = null;
-    let feedbackType = null;
-
-    await atomicReadModifyWrite((data) => {
-      // Check in team feedbacks first
-      const feedbackTeam = data.teams.find(t => t.id === feedbackTeamId);
-      let feedback = null;
-      if (feedbackTeam && feedbackTeam.teamFeedbacks) {
-        feedback = feedbackTeam.teamFeedbacks.find(f => f.id === feedbackId);
-      }
-      // Fall back to orphaned feedbacks from deleted teams
-      if (!feedback && Array.isArray(data.orphanedFeedbacks)) {
-        feedback = data.orphanedFeedbacks.find(f => f.id === feedbackId);
-      }
-      if (!feedback) return null;
-
-      feedbackTitle = feedback.title;
-      feedbackType = feedback.type;
-
-      if (!feedback.comments) {
-        feedback.comments = [];
-      }
-      feedback.comments.push(newComment);
-      return data;
-    });
-
-    // Notify super admin about new comment from team
-    if (feedbackTitle && smtpEnabled && mailer) {
-      const settings = await loadGlobalSettings();
-      const adminEmail = settings.adminEmail;
-
-      if (adminEmail) {
-        const typeLabel = feedbackType === 'bug' ? 'Bug Report' : 'Feature Request';
-        const safeFeedbackTitle = escapeHtml(feedbackTitle);
-        const safeTeamName = escapeHtml(requestingTeamName);
-        const safeAuthorName = escapeHtml(authorName);
-        const safeContent = escapeHtml(content.trim().slice(0, 1000));
-
-        try {
-          await mailer.sendMail({
-            from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-            to: adminEmail,
-            subject: `💬 New Comment on ${typeLabel}: ${feedbackTitle}`,
-            text: `New comment from ${requestingTeamName}
-
-${typeLabel}: ${feedbackTitle}
-Comment by: ${authorName}
-
-"${content.trim().slice(0, 1000)}"
-
----
-Log in to the Super Admin Dashboard to view and respond.
-`,
-            html: `
-<div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-  <h2 style="color: #4f46e5;">💬 New Comment</h2>
-  <p>New comment from <strong>${safeTeamName}</strong></p>
-  <div style="background: #f8fafc; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <p style="margin: 0 0 4px 0; color: #64748b; font-size: 14px;">${typeLabel}</p>
-    <h3 style="margin: 0; color: #1e293b;">${safeFeedbackTitle}</h3>
-  </div>
-  <div style="background: #f0f9ff; border: 1px solid #bae6fd; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <p style="margin: 0 0 8px 0; color: #0369a1; font-size: 14px;">Comment by <strong>${safeAuthorName}</strong>:</p>
-    <p style="margin: 0; color: #0c4a6e; white-space: pre-wrap;">${safeContent}</p>
-  </div>
-  <hr style="border: none; border-top: 1px solid #e2e8f0; margin: 24px 0;">
-  <p style="color: #94a3b8; font-size: 12px;">Log in to the Super Admin Dashboard to view and respond.</p>
-</div>
-`
-          });
-          addServerLog('info', 'email', `Comment notification sent to admin for: ${feedbackTitle}`);
-        } catch (emailErr) {
-          console.error('[Server] Failed to send comment notification to admin', emailErr);
-        }
-      }
-    }
-
-    res.json({ success: true, comment: newComment });
-  } catch (err) {
-    console.error('[Server] Failed to add comment', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-// Delete a comment from a feedback (only the author's team can delete)
-app.post('/api/feedbacks/comment/delete', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId, password, feedbackTeamId, feedbackId, commentId } = req.body || {};
-
-    // Authenticate the requesting team
-    const { error } = await authenticateTeam(teamId, password);
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!feedbackTeamId || !feedbackId || !commentId) {
-      return res.status(400).json({ error: 'missing_comment_data' });
-    }
-
-    await atomicReadModifyWrite((data) => {
-      // Check in team feedbacks first
-      const feedbackTeam = data.teams.find(t => t.id === feedbackTeamId);
-      let feedback = null;
-      if (feedbackTeam && feedbackTeam.teamFeedbacks) {
-        feedback = feedbackTeam.teamFeedbacks.find(f => f.id === feedbackId);
-      }
-      // Fall back to orphaned feedbacks from deleted teams
-      if (!feedback && Array.isArray(data.orphanedFeedbacks)) {
-        feedback = data.orphanedFeedbacks.find(f => f.id === feedbackId);
-      }
-      if (!feedback || !feedback.comments) return null;
-
-      // Find the comment and verify ownership
-      const comment = feedback.comments.find(c => c.id === commentId);
-      if (!comment) return null;
-
-      // Only allow deletion by the team that created the comment
-      if (comment.teamId !== teamId) {
-        return null; // Not authorized
-      }
-
-      feedback.comments = feedback.comments.filter(c => c.id !== commentId);
-      return data;
-    });
-
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to delete comment', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-// Delete a feedback (only the author's team can delete)
-app.post('/api/feedbacks/delete', teamWriteLimiter, async (req, res) => {
-  try {
-    const { teamId, password, feedbackId } = req.body || {};
-
-    // Authenticate the requesting team
-    const { team, error } = await authenticateTeam(teamId, password);
-    if (error) {
-      return res.status(401).json({ error });
-    }
-
-    if (!feedbackId) {
-      return res.status(400).json({ error: 'missing_feedback_id' });
-    }
-
-    let feedbackTitle = null;
-    let feedbackType = null;
-    const teamName = team ? team.name : 'Unknown Team';
-
-    await atomicReadModifyWrite((data) => {
-      const feedbackTeam = data.teams.find(t => t.id === teamId);
-      if (!feedbackTeam || !feedbackTeam.teamFeedbacks) return null;
-
-      // Find feedback and verify ownership
-      const feedback = feedbackTeam.teamFeedbacks.find(f => f.id === feedbackId);
-      if (!feedback || feedback.teamId !== teamId) {
-        return null; // Not authorized
-      }
-
-      feedbackTitle = feedback.title;
-      feedbackType = feedback.type;
-
-      feedbackTeam.teamFeedbacks = feedbackTeam.teamFeedbacks.filter(f => f.id !== feedbackId);
-      return data;
-    });
-
-    // Notify super admin about feedback deletion
-    if (feedbackTitle && smtpEnabled && mailer) {
-      const settings = await loadGlobalSettings();
-      const adminEmail = settings.adminEmail;
-
-      if (adminEmail) {
-        const typeLabel = feedbackType === 'bug' ? 'Bug Report' : 'Feature Request';
-        const safeFeedbackTitle = escapeHtml(feedbackTitle);
-        const safeTeamName = escapeHtml(teamName);
-
-        try {
-          await mailer.sendMail({
-            from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-            to: adminEmail,
-            subject: `🗑️ Feedback Deleted by Team: ${feedbackTitle}`,
-            text: `A feedback has been deleted by its author.
-
-${typeLabel}: ${feedbackTitle}
-Team: ${teamName}
-
----
-This notification was sent from RetroGemini.
-`,
-            html: `
-<div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-  <h2 style="color: #64748b;">🗑️ Feedback Deleted</h2>
-  <p>A feedback has been deleted by its author.</p>
-  <div style="background: #f8fafc; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <p style="margin: 0 0 4px 0; color: #64748b; font-size: 14px;">${typeLabel}</p>
-    <h3 style="margin: 0; color: #1e293b;">${safeFeedbackTitle}</h3>
-    <p style="margin: 8px 0 0 0; color: #64748b; font-size: 14px;">Team: ${safeTeamName}</p>
-  </div>
-  <hr style="border: none; border-top: 1px solid #e2e8f0; margin: 24px 0;">
-  <p style="color: #94a3b8; font-size: 12px;">This notification was sent from RetroGemini.</p>
-</div>
-`
-          });
-          addServerLog('info', 'email', `Feedback deletion notification sent to admin for: ${feedbackTitle}`);
-        } catch (emailErr) {
-          console.error('[Server] Failed to send feedback deletion notification to admin', emailErr);
-        }
-      }
-    }
-
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to delete feedback', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-app.post('/api/send-password-reset', async (req, res) => {
-  if (!smtpEnabled || !mailer) {
-    return res.status(501).json({ error: 'email_not_configured' });
-  }
-
-  const { email, teamName, resetLink, resetBaseUrl } = req.body || {};
-  const requestedLink = resetBaseUrl || resetLink;
-  if (!email || !requestedLink || !teamName) {
-    return res.status(400).json({ error: 'missing_fields' });
-  }
-
-  const safeTeamName = escapeHtml(teamName);
-  const safeResetLink = sanitizeEmailLink(requestedLink);
-  const safeResetUrl = new URL(safeResetLink);
-  const normalizedEmail = email.trim().toLowerCase();
-
-  try {
-    const currentData = await loadPersistedData();
-    const team = currentData.teams.find(
-      (t) => t.name.toLowerCase() === teamName.toLowerCase()
-    );
-    const facilitatorEmail = team?.facilitatorEmail?.trim().toLowerCase();
-
-    if (!team || !facilitatorEmail || facilitatorEmail !== normalizedEmail) {
-      return res.status(204).end();
-    }
-
-    const token = randomBytes(32).toString('hex');
-    const tokenHash = hashResetToken(token);
-    const now = Date.now();
-    const expiresAt = now + RESET_TOKEN_TTL_MS;
-
-    await atomicReadModifyWrite((data) => {
-      const tokens = pruneResetTokens(data.resetTokens);
-      const filtered = tokens.filter((entry) => entry.teamId !== team.id);
-      filtered.push({
-        tokenHash,
-        teamId: team.id,
-        createdAt: now,
-        expiresAt
-      });
-      data.resetTokens = filtered;
-      return data;
-    });
-
-    safeResetUrl.searchParams.set('reset', token);
-    const resetLinkWithToken = safeResetUrl.toString();
-    const safeResetLinkHtml = escapeHtml(resetLinkWithToken);
-
-    await mailer.sendMail({
-      from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-      to: email,
-      subject: `Password Reset - ${teamName}`,
-      text: `Hello,
-
-You have requested a password reset for the team "${teamName}".
-
-Click this link to reset your password: ${resetLinkWithToken}
-
-This link is valid for 1 hour.
-
-If you did not request this reset, please ignore this email.
-`,
-      html: `<p>Hello,</p>
-<p>You have requested a password reset for the team <strong>${safeTeamName}</strong>.</p>
-<p><a href="${safeResetLinkHtml}" target="_blank" rel="noreferrer">Click here to reset your password</a></p>
-<p>This link is valid for 1 hour.</p>
-<p><em>If you did not request this reset, please ignore this email.</em></p>`
-    });
-
-    res.status(204).end();
-  } catch (err) {
-    console.error('[Server] Failed to send password reset email', err);
-    res.status(500).json({ error: 'send_failed' });
-  }
-});
-
-app.post('/api/password-reset/verify', async (req, res) => {
-  const { token } = req.body || {};
-  if (!token) {
-    return res.status(400).json({ error: 'missing_token' });
-  }
-
-  try {
-    const currentData = await loadPersistedData();
-    const prunedTokens = pruneResetTokens(currentData.resetTokens);
-    const tokenHash = hashResetToken(token);
-    const tokenEntry = prunedTokens.find((entry) => entry.tokenHash === tokenHash);
-
-    if (prunedTokens.length !== currentData.resetTokens.length) {
-      await atomicReadModifyWrite((data) => {
-        data.resetTokens = prunedTokens;
-        return data;
-      });
-    }
-
-    if (!tokenEntry) {
-      return res.json({ valid: false });
-    }
-
-    const team = currentData.teams.find((t) => t.id === tokenEntry.teamId);
-    if (!team) {
-      return res.json({ valid: false });
-    }
-
-    return res.json({ valid: true, teamName: team.name });
-  } catch (err) {
-    console.error('[Server] Failed to verify reset token', err);
-    return res.status(500).json({ error: 'verification_failed' });
-  }
-});
-
-app.post('/api/password-reset/confirm', async (req, res) => {
-  const { token, newPassword } = req.body || {};
-  if (!token || !newPassword) {
-    return res.status(400).json({ error: 'missing_fields' });
-  }
-
-  if (newPassword.length < 4) {
-    return res.status(400).json({ error: 'password_too_short' });
-  }
-
-  let updated = false;
-  let teamName = null;
-
-  try {
-    await atomicReadModifyWrite((data) => {
-      data.resetTokens = pruneResetTokens(data.resetTokens);
-      const tokenHash = hashResetToken(token);
-      const tokenIndex = data.resetTokens.findIndex((entry) => entry.tokenHash === tokenHash);
-      if (tokenIndex === -1) {
-        return null;
-      }
-      const tokenEntry = data.resetTokens[tokenIndex];
-      const team = data.teams.find((t) => t.id === tokenEntry.teamId);
-      if (!team) {
-        return null;
-      }
-      team.passwordHash = newPassword;
-      teamName = team.name;
-      data.resetTokens.splice(tokenIndex, 1);
-      updated = true;
-      return data;
-    });
-
-    if (!updated) {
-      return res.status(400).json({ error: 'invalid_or_expired_token' });
-    }
-
-    return res.json({
-      success: true,
-      message: `Password updated for ${teamName}. You can now log in.`,
-      teamName
-    });
-  } catch (err) {
-    console.error('[Server] Failed to reset password', err);
-    return res.status(500).json({ error: 'reset_failed' });
-  }
-});
-
-// Super Admin endpoints
-// Set SUPER_ADMIN_PASSWORD environment variable to enable super admin access
-const SUPER_ADMIN_PASSWORD = process.env.SUPER_ADMIN_PASSWORD;
-
-/**
- * Constant-time string comparison to prevent timing attacks
- * @param {string} a - First string
- * @param {string} b - Second string
- * @returns {boolean} - True if strings are equal
- */
-const secureCompare = (a, b) => {
-  if (typeof a !== 'string' || typeof b !== 'string') {
-    return false;
-  }
-
-  // Convert strings to buffers of equal length to prevent length-based timing attacks
-  const bufferA = Buffer.from(a);
-  const bufferB = Buffer.from(b);
-
-  // If lengths differ, compare against a dummy buffer to maintain constant time
-  if (bufferA.length !== bufferB.length) {
-    // Still perform a comparison to avoid timing leak on length check
-    timingSafeEqual(bufferA, bufferA);
-    return false;
-  }
-
-  return timingSafeEqual(bufferA, bufferB);
-};
-
-app.post('/api/super-admin/verify', authLimiter, (req, res) => {
-  const { password } = req.body || {};
-
-  if (!SUPER_ADMIN_PASSWORD) {
-    return res.status(503).json({ error: 'super_admin_not_configured' });
-  }
-
-  if (secureCompare(password, SUPER_ADMIN_PASSWORD)) {
-    // Generate session token for browser refresh persistence
-    const sessionToken = createSuperAdminToken();
-    return res.json({ success: true, sessionToken });
-  }
-
-  return res.status(401).json({ error: 'invalid_password' });
-});
-
-// Validate super admin session token (for browser refresh)
-app.post('/api/super-admin/validate-session', authLimiter, (req, res) => {
-  const { sessionToken } = req.body || {};
-
-  if (!SUPER_ADMIN_PASSWORD) {
-    return res.status(503).json({ error: 'super_admin_not_configured' });
-  }
-
-  if (!sessionToken || !validateSuperAdminToken(sessionToken)) {
-    return res.status(401).json({ error: 'invalid_or_expired_token' });
-  }
-
-  return res.json({ success: true });
-});
-
-app.post('/api/super-admin/teams', superAdminActionLimiter, (req, res) => {
-  const { password } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  // Return only team metadata (no session content for privacy)
-  refreshPersistedData()
-    .then((currentData) => {
-      const sanitizedTeams = currentData.teams.map((t) => ({
-        id: t.id,
-        name: t.name,
-        facilitatorEmail: t.facilitatorEmail,
-        members: (t.members || []).map((m) => ({ id: m.id, name: m.name, color: m.color, role: m.role })),
-        lastConnectionDate: t.lastConnectionDate
-      }));
-      res.json({ teams: sanitizedTeams });
-    })
-    .catch((err) => {
-      console.error('[Server] Failed to load persisted data', err);
-      res.status(500).json({ error: 'failed_to_load' });
-    });
-});
-
-app.post('/api/super-admin/feedbacks', superAdminActionLimiter, (req, res) => {
-  const { password } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  refreshPersistedData()
-    .then((currentData) => {
-      const feedbacks = currentData.teams.flatMap((team) =>
-        (team.teamFeedbacks || []).map((feedback) => ({
-          ...feedback,
-          teamId: feedback.teamId || team.id,
-          teamName: feedback.teamName || team.name,
-          isRead: feedback.isRead ?? false,
-          status: feedback.status || 'pending'
-        }))
-      );
-      // Include orphaned feedbacks from deleted teams
-      const orphaned = (currentData.orphanedFeedbacks || []).map((feedback) => ({
-        ...feedback,
-        isRead: feedback.isRead ?? false,
-        status: feedback.status || 'pending'
-      }));
-      feedbacks.push(...orphaned);
-      feedbacks.sort((a, b) => new Date(b.submittedAt).getTime() - new Date(a.submittedAt).getTime());
-      res.json({ feedbacks });
-    })
-    .catch((err) => {
-      console.error('[Server] Failed to load feedbacks', err);
-      res.status(500).json({ error: 'failed_to_load' });
-    });
-});
-
-app.post('/api/super-admin/update-email', superAdminActionLimiter, async (req, res) => {
-  const { password, teamId, facilitatorEmail } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  if (!teamId) {
-    return res.status(400).json({ error: 'missing_team_id' });
-  }
-
-  try {
-    await atomicReadModifyWrite((data) => {
-      const team = data.teams.find(t => t.id === teamId);
-      if (!team) return null;
-      team.facilitatorEmail = facilitatorEmail || undefined;
-      return data;
-    });
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to update email', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-app.post('/api/super-admin/feedbacks/update', superAdminActionLimiter, async (req, res) => {
-  const { password, teamId, feedbackId, updates } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  if (!teamId || !feedbackId || !updates) {
-    return res.status(400).json({ error: 'missing_feedback_data' });
-  }
-
-  try {
-    let statusChanged = false;
-    let oldStatus = null;
-    let newStatus = null;
-    let feedbackTitle = null;
-    let feedbackType = null;
-    let teamEmail = null;
-    let teamName = null;
-
-    await atomicReadModifyWrite((data) => {
-      // Check in team feedbacks first
-      const team = data.teams.find(t => t.id === teamId);
-      let feedback = null;
-      if (team && team.teamFeedbacks) {
-        feedback = team.teamFeedbacks.find(f => f.id === feedbackId);
-      }
-      // Fall back to orphaned feedbacks from deleted teams
-      if (!feedback && Array.isArray(data.orphanedFeedbacks)) {
-        feedback = data.orphanedFeedbacks.find(f => f.id === feedbackId);
-      }
-      if (!feedback) return null;
-
-      // Check if status is changing
-      if (updates.status && updates.status !== feedback.status) {
-        statusChanged = true;
-        oldStatus = feedback.status;
-        newStatus = updates.status;
-        feedbackTitle = feedback.title;
-        feedbackType = feedback.type;
-        teamEmail = team ? team.facilitatorEmail : undefined;
-        teamName = team ? team.name : feedback.teamName;
-      }
-
-      Object.assign(feedback, updates);
-      if (!feedback.teamName) {
-        feedback.teamName = team ? team.name : 'Deleted Team';
-      }
-      if (!feedback.teamId) {
-        feedback.teamId = team ? team.id : teamId;
-      }
-      return data;
-    });
-
-    // Send email notification if status changed and team has an email configured
-    if (statusChanged && teamEmail && smtpEnabled && mailer) {
-      const statusLabels = {
-        pending: 'Pending',
-        in_progress: 'In Progress',
-        resolved: 'Resolved',
-        rejected: 'Rejected'
-      };
-      const statusEmojis = {
-        pending: '⏳',
-        in_progress: '🔄',
-        resolved: '✅',
-        rejected: '❌'
-      };
-      const typeLabel = feedbackType === 'bug' ? 'Bug Report' : 'Feature Request';
-      const safeFeedbackTitle = escapeHtml(feedbackTitle);
-      const safeTeamName = escapeHtml(teamName);
-
-      try {
-        await mailer.sendMail({
-          from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-          to: teamEmail,
-          subject: `${statusEmojis[newStatus]} Feedback Status Updated: ${feedbackTitle}`,
-          text: `Hello ${teamName},
-
-The status of your ${typeLabel} has been updated.
-
-Title: ${feedbackTitle}
-Previous Status: ${statusLabels[oldStatus]}
-New Status: ${statusLabels[newStatus]}
-
----
-This notification was sent because your feedback status was updated in RetroGemini.
-`,
-          html: `
-<div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-  <h2 style="color: #4f46e5;">
-    ${statusEmojis[newStatus]} Feedback Status Updated
-  </h2>
-  <p>Hello <strong>${safeTeamName}</strong>,</p>
-  <p>The status of your ${typeLabel} has been updated.</p>
-  <div style="background: #f8fafc; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <h3 style="margin: 0 0 8px 0; color: #1e293b;">${safeFeedbackTitle}</h3>
-    <p style="margin: 4px 0; color: #64748b; font-size: 14px;">
-      <strong>Previous Status:</strong> <span style="color: #94a3b8;">${statusLabels[oldStatus]}</span><br>
-      <strong>New Status:</strong> <span style="color: ${newStatus === 'resolved' ? '#16a34a' : newStatus === 'rejected' ? '#dc2626' : newStatus === 'in_progress' ? '#2563eb' : '#ca8a04'}; font-weight: bold;">${statusLabels[newStatus]}</span>
-    </p>
-  </div>
-  <hr style="border: none; border-top: 1px solid #e2e8f0; margin: 24px 0;">
-  <p style="color: #94a3b8; font-size: 12px;">
-    This notification was sent because your feedback status was updated in RetroGemini.
-  </p>
-</div>
-`
-        });
-        addServerLog('info', 'email', `Feedback status notification sent to ${teamEmail} for: ${feedbackTitle}`);
-      } catch (emailErr) {
-        // Log but don't fail - the update was successful
-        console.error('[Server] Failed to send feedback status notification email', emailErr);
-        addServerLog('warn', 'email', `Failed to send feedback status notification to ${teamEmail}: ${emailErr.message}`);
-      }
-    }
-
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to update feedback', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-app.post('/api/super-admin/feedbacks/delete', superAdminActionLimiter, async (req, res) => {
-  const { password, teamId, feedbackId } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  if (!teamId || !feedbackId) {
-    return res.status(400).json({ error: 'missing_feedback_data' });
-  }
-
-  try {
-    let feedbackTitle = null;
-    let feedbackType = null;
-    let teamEmail = null;
-    let teamName = null;
-
-    await atomicReadModifyWrite((data) => {
-      // Check in team feedbacks first
-      const team = data.teams.find(t => t.id === teamId);
-      let found = false;
-      if (team && team.teamFeedbacks) {
-        const feedback = team.teamFeedbacks.find(f => f.id === feedbackId);
-        if (feedback) {
-          found = true;
-          feedbackTitle = feedback.title;
-          feedbackType = feedback.type;
-          teamEmail = team.facilitatorEmail;
-          teamName = team.name;
-          team.teamFeedbacks = team.teamFeedbacks.filter(f => f.id !== feedbackId);
-        }
-      }
-      // Fall back to orphaned feedbacks from deleted teams
-      if (!found && Array.isArray(data.orphanedFeedbacks)) {
-        const feedback = data.orphanedFeedbacks.find(f => f.id === feedbackId);
-        if (feedback) {
-          feedbackTitle = feedback.title;
-          feedbackType = feedback.type;
-          teamName = feedback.teamName;
-          data.orphanedFeedbacks = data.orphanedFeedbacks.filter(f => f.id !== feedbackId);
-          found = true;
-        }
-      }
-      if (!found) return null;
-      return data;
-    });
-
-    // Send notification to team about deletion
-    if (feedbackTitle && teamEmail && smtpEnabled && mailer) {
-      const typeLabel = feedbackType === 'bug' ? 'Bug Report' : 'Feature Request';
-      const safeFeedbackTitle = escapeHtml(feedbackTitle);
-      const safeTeamName = escapeHtml(teamName);
-
-      try {
-        await mailer.sendMail({
-          from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-          to: teamEmail,
-          subject: `🗑️ Feedback Deleted: ${feedbackTitle}`,
-          text: `Hello ${teamName},
-
-Your ${typeLabel} "${feedbackTitle}" has been deleted by the administrator.
-
-If you have questions about this action, please contact the administrator.
-
----
-This notification was sent from RetroGemini.
-`,
-          html: `
-<div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-  <h2 style="color: #dc2626;">🗑️ Feedback Deleted</h2>
-  <p>Hello <strong>${safeTeamName}</strong>,</p>
-  <p>Your ${typeLabel} "<strong>${safeFeedbackTitle}</strong>" has been deleted by the administrator.</p>
-  <p>If you have questions about this action, please contact the administrator.</p>
-  <hr style="border: none; border-top: 1px solid #e2e8f0; margin: 24px 0;">
-  <p style="color: #94a3b8; font-size: 12px;">This notification was sent from RetroGemini.</p>
-</div>
-`
-        });
-        addServerLog('info', 'email', `Feedback deletion notification sent to ${teamEmail} for: ${feedbackTitle}`);
-      } catch (emailErr) {
-        console.error('[Server] Failed to send feedback deletion notification', emailErr);
-      }
-    }
-
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to delete feedback', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-// Super admin adds a comment to a feedback
-app.post('/api/super-admin/feedbacks/comment', superAdminActionLimiter, async (req, res) => {
-  const { password, teamId, feedbackId, content } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  if (!teamId || !feedbackId || !content) {
-    return res.status(400).json({ error: 'missing_comment_data' });
-  }
-
-  try {
-    let feedbackTitle = null;
-    let feedbackType = null;
-    let teamEmail = null;
-    let teamName = null;
-
-    const commentId = `comment_${Date.now()}_${Math.random().toString(36).substr(2, 5)}`;
-    const newComment = {
-      id: commentId,
-      feedbackId,
-      teamId: 'super-admin',
-      teamName: 'Super Admin',
-      authorId: 'super-admin',
-      authorName: 'Super Admin',
-      content: content.trim().slice(0, 1000),
-      createdAt: new Date().toISOString(),
-      isAdmin: true
-    };
-
-    await atomicReadModifyWrite((data) => {
-      // Check in team feedbacks first
-      const team = data.teams.find(t => t.id === teamId);
-      let feedback = null;
-      if (team && team.teamFeedbacks) {
-        feedback = team.teamFeedbacks.find(f => f.id === feedbackId);
-      }
-      // Fall back to orphaned feedbacks from deleted teams
-      if (!feedback && Array.isArray(data.orphanedFeedbacks)) {
-        feedback = data.orphanedFeedbacks.find(f => f.id === feedbackId);
-      }
-      if (!feedback) return null;
-
-      feedbackTitle = feedback.title;
-      feedbackType = feedback.type;
-      teamEmail = team ? team.facilitatorEmail : undefined;
-      teamName = team ? team.name : feedback.teamName;
-
-      if (!feedback.comments) {
-        feedback.comments = [];
-      }
-      feedback.comments.push(newComment);
-      return data;
-    });
-
-    // Send notification to team about new admin comment
-    if (feedbackTitle && teamEmail && smtpEnabled && mailer) {
-      const typeLabel = feedbackType === 'bug' ? 'Bug Report' : 'Feature Request';
-      const safeFeedbackTitle = escapeHtml(feedbackTitle);
-      const safeTeamName = escapeHtml(teamName);
-      const safeContent = escapeHtml(content.trim().slice(0, 1000));
-
-      try {
-        await mailer.sendMail({
-          from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-          to: teamEmail,
-          subject: `💬 Admin Comment on: ${feedbackTitle}`,
-          text: `Hello ${teamName},
-
-The administrator has added a comment on your ${typeLabel} "${feedbackTitle}":
-
-"${content.trim().slice(0, 1000)}"
-
----
-This notification was sent from RetroGemini.
-`,
-          html: `
-<div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-  <h2 style="color: #4f46e5;">💬 Admin Comment</h2>
-  <p>Hello <strong>${safeTeamName}</strong>,</p>
-  <p>The administrator has added a comment on your ${typeLabel}:</p>
-  <div style="background: #f8fafc; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <h3 style="margin: 0 0 8px 0; color: #1e293b;">${safeFeedbackTitle}</h3>
-  </div>
-  <div style="background: #fffbeb; border: 1px solid #fcd34d; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <p style="margin: 0; color: #78350f; white-space: pre-wrap;">${safeContent}</p>
-  </div>
-  <hr style="border: none; border-top: 1px solid #e2e8f0; margin: 24px 0;">
-  <p style="color: #94a3b8; font-size: 12px;">This notification was sent from RetroGemini.</p>
-</div>
-`
-        });
-        addServerLog('info', 'email', `Admin comment notification sent to ${teamEmail} for: ${feedbackTitle}`);
-      } catch (emailErr) {
-        console.error('[Server] Failed to send admin comment notification', emailErr);
-      }
-    }
-
-    res.json({ success: true, comment: newComment });
-  } catch (err) {
-    console.error('[Server] Failed to add admin comment', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-app.post('/api/super-admin/update-password', superAdminActionLimiter, async (req, res) => {
-  const { password, teamId, newPassword } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  if (!teamId) {
-    return res.status(400).json({ error: 'missing_team_id' });
-  }
-
-  if (!newPassword || newPassword.length < 4) {
-    return res.status(400).json({ error: 'password_too_short' });
-  }
-
-  try {
-    await atomicReadModifyWrite((data) => {
-      const team = data.teams.find(t => t.id === teamId);
-      if (!team) return null;
-      team.passwordHash = newPassword;
-      return data;
-    });
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to update password', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-app.post('/api/super-admin/rename-team', superAdminActionLimiter, async (req, res) => {
-  const { password, teamId, newName } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  if (!teamId) {
-    return res.status(400).json({ error: 'missing_team_id' });
-  }
-
-  if (!newName || newName.trim().length === 0) {
-    return res.status(400).json({ error: 'team_name_empty' });
-  }
-
-  const trimmedName = newName.trim();
-
-  try {
-    await atomicReadModifyWrite((data) => {
-      const team = data.teams.find(t => t.id === teamId);
-      if (!team) return null;
-
-      // Check if another team already has this name (case-insensitive)
-      const existingTeam = data.teams.find(t => t.id !== teamId && t.name.toLowerCase() === trimmedName.toLowerCase());
-      if (existingTeam) return null;
-
-      team.name = trimmedName;
-      return data;
-    });
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to rename team', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-app.post(
-  '/api/super-admin/restore',
-  superAdminActionLimiter,
-  express.raw({
-    type: ['application/gzip', 'application/x-gzip', 'application/octet-stream', 'application/json'],
-    limit: '1gb'
-  }),
-  async (req, res) => {
-    const password = req.header('x-super-admin-password');
-    const sessionToken = req.header('x-super-admin-session-token');
-
-    if (!validateSuperAdminAuth({ password, sessionToken })) {
-      return res.status(401).json({ error: 'unauthorized' });
-    }
-
-    if (!req.body || !(req.body instanceof Buffer) || req.body.length === 0) {
-      return res.status(400).json({ error: 'missing_archive' });
-    }
-
-    try {
-      let data;
-
-      // Try to decompress as gzip first
-      try {
-        const decompressed = gunzipSync(req.body);
-        data = JSON.parse(decompressed.toString('utf8'));
-      } catch {
-        // If gzip fails, try parsing as plain JSON
-        try {
-          data = JSON.parse(req.body.toString('utf8'));
-        } catch {
-          return res.status(400).json({ error: 'invalid_backup_format' });
-        }
-      }
-
-      // Validate the data structure
-      if (!data || typeof data !== 'object') {
-        return res.status(400).json({ error: 'invalid_backup_data' });
-      }
-
-      // Ensure teams array exists
-      if (!Array.isArray(data.teams)) {
-        data.teams = [];
-      }
-
-      // Save to database (works for both SQLite and PostgreSQL)
-      persistedData = await savePersistedData(data);
-
-      const teamCount = data.teams.length;
-      console.info(`[Server] Restored backup: ${teamCount} team(s)`);
-
-      res.json({ success: true, teamsRestored: teamCount });
-    } catch (err) {
-      console.error('[Server] Failed to restore backup', err);
-      res.status(500).json({ error: 'restore_failed' });
-    }
-  }
-);
-
-app.post('/api/super-admin/backup', superAdminActionLimiter, (req, res) => {
-  const { password } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  refreshPersistedData()
-    .then((currentData) => {
-      const timestamp = new Date().toISOString().replace(/[:.]/g, '-');
-      const filename = `retrogemini-backup-${timestamp}.json.gz`;
-
-      // Export data as gzipped JSON (works for both SQLite and PostgreSQL)
-      const jsonData = JSON.stringify(currentData, null, 2);
-      const compressed = gzipSync(Buffer.from(jsonData, 'utf8'));
-
-      const teamCount = currentData.teams?.length || 0;
-      console.info(`[Server] Creating backup: ${teamCount} team(s)`);
-
-      res.setHeader('Content-Type', 'application/gzip');
-      res.setHeader('Content-Disposition', `attachment; filename="${filename}"`);
-      res.setHeader('Cache-Control', 'no-store');
-      res.send(compressed);
-    })
-    .catch((err) => {
-      console.error('[Server] Failed to create backup', err);
-      res.status(500).json({ error: 'backup_failed' });
-    });
-});
-
-// Public endpoint to get info message
-app.get('/api/info-message', async (_req, res) => {
-  try {
-    const settings = await loadGlobalSettings();
-    res.json({ infoMessage: settings.infoMessage || '' });
-  } catch (err) {
-    console.error('[Server] Failed to load info message', err);
-    res.status(500).json({ error: 'failed_to_load' });
-  }
-});
-
-// Super admin endpoint to update info message
-app.post('/api/super-admin/info-message', superAdminActionLimiter, async (req, res) => {
-  const { password, infoMessage } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  try {
-    const settings = await loadGlobalSettings();
-    settings.infoMessage = infoMessage || '';
-    await saveGlobalSettings(settings);
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to update info message', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-// Super admin endpoint to get admin email
-app.post('/api/super-admin/admin-email', superAdminActionLimiter, async (req, res) => {
-  const { password } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  try {
-    const settings = await loadGlobalSettings();
-    res.json({ adminEmail: settings.adminEmail || '', notifyNewTeam: !!settings.notifyNewTeam });
-  } catch (err) {
-    console.error('[Server] Failed to load admin email', err);
-    res.status(500).json({ error: 'failed_to_load' });
-  }
-});
-
-// Super admin endpoint to update admin email
-app.post('/api/super-admin/update-admin-email', superAdminActionLimiter, async (req, res) => {
-  const { password, adminEmail } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  try {
-    const settings = await loadGlobalSettings();
-    settings.adminEmail = adminEmail || '';
-    await saveGlobalSettings(settings);
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to update admin email', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-// Super admin endpoint to toggle new team creation email notification
-app.post('/api/super-admin/update-notify-new-team', superAdminActionLimiter, async (req, res) => {
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  try {
-    const { notifyNewTeam } = req.body || {};
-    const settings = await loadGlobalSettings();
-    settings.notifyNewTeam = !!notifyNewTeam;
-    await saveGlobalSettings(settings);
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to update notify new team setting', err);
-    res.status(500).json({ error: 'failed_to_save' });
-  }
-});
-
-// Endpoint to send feedback notification email to admin
-app.post('/api/super-admin/notify-feedback', superAdminActionLimiter, async (req, res) => {
-  const { password, feedback } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  if (!smtpEnabled || !mailer) {
-    return res.status(501).json({ error: 'email_not_configured' });
-  }
-
-  try {
-    const settings = await loadGlobalSettings();
-    const adminEmail = settings.adminEmail;
-
-    if (!adminEmail) {
-      return res.status(400).json({ error: 'admin_email_not_configured' });
-    }
-
-    if (!feedback || !feedback.title || !feedback.type) {
-      return res.status(400).json({ error: 'missing_feedback_data' });
-    }
-
-    const typeLabel = feedback.type === 'bug' ? 'Bug Report' : 'Feature Request';
-    const typeEmoji = feedback.type === 'bug' ? '🐛' : '✨';
-    const safeFeedbackTitle = escapeHtml(feedback.title);
-    const safeFeedbackTeamName = escapeHtml(feedback.teamName);
-    const safeFeedbackSubmittedBy = escapeHtml(feedback.submittedByName);
-    const safeFeedbackDescription = escapeHtml(feedback.description);
-    const feedbackDate = new Date(feedback.submittedAt).toLocaleString();
-
-    await mailer.sendMail({
-      from: process.env.FROM_EMAIL || process.env.SMTP_USER,
-      to: adminEmail,
-      subject: `${typeEmoji} New ${typeLabel}: ${feedback.title}`,
-      text: `New ${typeLabel} submitted
-
-Title: ${feedback.title}
-Type: ${typeLabel}
-Team: ${feedback.teamName}
-Submitted by: ${feedback.submittedByName}
-Date: ${feedbackDate}
-
-Description:
-${feedback.description}
-
----
-Log in to the Super Admin Dashboard to review and respond to this feedback.
-`,
-      html: `
-<div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto;">
-  <h2 style="color: ${feedback.type === 'bug' ? '#dc2626' : '#7c3aed'};">
-    ${typeEmoji} New ${typeLabel}
-  </h2>
-  <div style="background: #f8fafc; border-radius: 8px; padding: 16px; margin: 16px 0;">
-    <h3 style="margin: 0 0 8px 0; color: #1e293b;">${safeFeedbackTitle}</h3>
-    <p style="margin: 4px 0; color: #64748b; font-size: 14px;">
-      <strong>Team:</strong> ${safeFeedbackTeamName}<br>
-      <strong>Submitted by:</strong> ${safeFeedbackSubmittedBy}<br>
-      <strong>Date:</strong> ${feedbackDate}
-    </p>
-  </div>
-  <div style="margin: 16px 0;">
-    <h4 style="color: #475569; margin-bottom: 8px;">Description:</h4>
-    <p style="color: #334155; white-space: pre-wrap;">${safeFeedbackDescription}</p>
-  </div>
-  ${feedback.images && feedback.images.length > 0 ? `
-  <p style="color: #64748b; font-size: 14px;">
-    <em>${feedback.images.length} image(s) attached - view in Super Admin Dashboard</em>
-  </p>
-  ` : ''}
-  <hr style="border: none; border-top: 1px solid #e2e8f0; margin: 24px 0;">
-  <p style="color: #94a3b8; font-size: 12px;">
-    Log in to the Super Admin Dashboard to review and respond to this feedback.
-  </p>
-</div>
-`
-    });
-
-    addServerLog('info', 'email', `Feedback notification sent to ${adminEmail} for: ${feedback.title}`);
-    res.json({ success: true });
-  } catch (err) {
-    console.error('[Server] Failed to send feedback notification email', err);
-    res.status(500).json({ error: 'send_failed' });
-  }
-});
-
-// Dedicated rate limiter for live-session polling (more lenient than action limiter)
-const superAdminPollingLimiter = rateLimit({
-  windowMs: 60 * 1000,
-  max: 30,
-  message: { error: 'too_many_attempts', retryAfter: '1 minute' },
-  standardHeaders: true,
-  legacyHeaders: false
-});
-
-// Super admin endpoint to get active sessions (live monitoring)
-app.post('/api/super-admin/active-sessions', superAdminPollingLimiter, async (req, res) => {
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  try {
-    const activeSessions = [];
-
-    // Use fetchSockets() to discover sockets across ALL pods (works with
-    // Redis / PostgreSQL adapters).  io.sockets.adapter.rooms only returns
-    // rooms local to the current pod, so a session running on another pod
-    // would be invisible.
-    const allSockets = await io.fetchSockets();
-
-    // Group sockets by session room (skip each socket's personal room)
-    const sessionRooms = new Map();
-    for (const s of allSockets) {
-      if (!s.data.userId || !s.data.userName) continue;
-      for (const room of s.rooms) {
-        if (room === s.id) continue; // personal room
-        if (!sessionRooms.has(room)) sessionRooms.set(room, []);
-        sessionRooms.get(room).push({ id: s.data.userId, name: s.data.userName });
-      }
-    }
-
-    for (const [roomId, participants] of sessionRooms.entries()) {
-      // Try to get session data from cache or database
-      let sessionData = sessions.get(roomId);
-      if (!sessionData) {
-        sessionData = await loadSessionState(roomId);
-      }
-
-      // Determine session type and details
-      const isHealthCheck = sessionData && (sessionData.templateId || sessionData.dimensions);
-      let teamName = 'Unknown';
-
-      // Try to get team name from persisted data
-      if (sessionData?.teamId && persistedData.teams) {
-        const team = persistedData.teams.find(t => t.id === sessionData.teamId);
-        if (team) {
-          teamName = team.name;
-        }
-      }
-
-      const sessionInfo = {
-        sessionId: roomId,
-        type: isHealthCheck ? 'healthcheck' : 'retrospective',
-        teamId: sessionData?.teamId || '',
-        teamName,
-        sessionName: sessionData?.name || 'Unknown Session',
-        phase: sessionData?.phase || 'Unknown',
-        status: sessionData?.status || 'IN_PROGRESS',
-        participants,
-        connectedCount: participants.length
-      };
-
-      activeSessions.push(sessionInfo);
-    }
-
-    res.json({ sessions: activeSessions });
-  } catch (err) {
-    console.error('[Server] Failed to get active sessions', err);
-    res.status(500).json({ error: 'failed_to_load' });
-  }
-});
-
-// Super admin endpoint to get server logs
-app.post('/api/super-admin/logs', superAdminActionLimiter, async (req, res) => {
-  const { password, filter } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  try {
-    let logs = [...serverLogs];
-
-    // Apply filters if specified
-    if (filter) {
-      if (filter.level) {
-        logs = logs.filter(l => l.level === filter.level);
-      }
-      if (filter.source) {
-        logs = logs.filter(l => l.source === filter.source);
-      }
-    }
-
-    // Return logs in reverse order (newest first)
-    res.json({ logs: logs.reverse() });
-  } catch (err) {
-    console.error('[Server] Failed to get server logs', err);
-    res.status(500).json({ error: 'failed_to_load' });
-  }
-});
-
-// Super admin endpoint to clear server logs
-app.post('/api/super-admin/clear-logs', superAdminActionLimiter, async (req, res) => {
-  const { password } = req.body || {};
-
-  if (!validateSuperAdminAuth(req.body)) {
-    return res.status(401).json({ error: 'unauthorized' });
-  }
-
-  serverLogs.length = 0;
-  addServerLog('info', 'server', 'Server logs cleared by admin');
-  res.json({ success: true });
-});
-
-// Serve static files from dist folder
 app.use(express.static(join(__dirname, 'dist')));
 
-// In-memory cache for sessions (per pod)
-const sessions = new Map(); // sessionId -> session data
-
-const buildSessionRoster = async (sessionId) => {
-  try {
-    const sockets = await io.in(sessionId).fetchSockets();
-    return sockets
-      .map((connectedSocket) => ({
-        id: connectedSocket.data.userId,
-        name: connectedSocket.data.userName
-      }))
-      .filter((member) => member.id && member.name);
-  } catch (err) {
-    console.warn('[Server] Failed to fetch full roster across pods', err);
-    const localMembers = [];
-    for (const connectedSocket of io.sockets.sockets.values()) {
-      if (!connectedSocket.rooms.has(sessionId)) continue;
-      if (connectedSocket.data.userId && connectedSocket.data.userName) {
-        localMembers.push({
-          id: connectedSocket.data.userId,
-          name: connectedSocket.data.userName
-        });
-      }
-    }
-    return localMembers;
-  }
-};
-
-const leaveCurrentSession = async (socket) => {
-  const sessionId = socket.sessionId;
-  if (!sessionId) return;
-
-  console.log(`[Server] ${socket.userName || 'Unknown'} leaving session ${sessionId}`);
-  socket.leave(sessionId);
-
-  const room = io.sockets.adapter.rooms.get(sessionId);
-  console.log(`[Server] Session ${sessionId} now has ${room?.size || 0} connected clients`);
-
-  socket.to(sessionId).emit('member-left', {
-    userId: socket.userId,
-    userName: socket.userName
-  });
-
-  const roster = await buildSessionRoster(sessionId);
-  io.to(sessionId).emit('member-roster', roster);
-
-  socket.sessionId = null;
-};
-
-io.on('connection', (socket) => {
-  console.log('[Server] Client connected:', socket.id);
-
-  // Join a session room
-  socket.on('join-session', async ({ sessionId, userId, userName }) => {
-    console.log(`[Server] User ${userName} (${userId}) joining session ${sessionId}`);
-
-    // Leave any previously joined session to avoid cross-room events
-    if (socket.sessionId && socket.sessionId !== sessionId) {
-      await leaveCurrentSession(socket);
-    }
-
-    socket.join(sessionId);
-    socket.sessionId = sessionId;
-    socket.userId = userId;
-    socket.userName = userName;
-    socket.data.userId = userId;
-    socket.data.userName = userName;
-
-    // Share current roster (including the new joiner) with everyone in the room
-    const roster = await buildSessionRoster(sessionId);
-    io.to(sessionId).emit('member-roster', roster);
-
-    // Log current room members
-    const room = io.sockets.adapter.rooms.get(sessionId);
-    console.log(`[Server] Session ${sessionId} now has ${room?.size || 0} connected clients`);
-
-    // Send current session state to the new joiner
-    if (usePostgres || sqliteDb) {
-      const persistedSession = await loadSessionState(sessionId);
-      if (persistedSession) {
-        sessions.set(sessionId, persistedSession);
-        console.log(`[Server] Sending persisted session state to ${userName}`);
-        socket.emit('session-update', persistedSession);
-      } else if (sessions.has(sessionId)) {
-        console.log(`[Server] Sending cached session state to ${userName}`);
-        socket.emit('session-update', sessions.get(sessionId));
-      }
-    }
-
-    // Notify others that someone joined
-    socket.to(sessionId).emit('member-joined', { userId, userName });
-
-    // Update team lastConnectionDate when a non-facilitator joins a session
-    try {
-      const sessionData = await loadSessionState(sessionId) || sessions.get(sessionId);
-      if (sessionData?.teamId) {
-        // Retry up to 3 times in case of revision conflicts
-        for (let attempt = 0; attempt < 3; attempt++) {
-          const currentData = await loadPersistedData();
-          const team = currentData.teams.find(t => t.id === sessionData.teamId);
-          if (!team) break;
-          const member = team.members.find(m => m.id === userId);
-          if (!member || member.role === 'facilitator') break;
-          team.lastConnectionDate = new Date().toISOString();
-          const revision = Number(currentData.meta?.revision ?? 0);
-          const result = await atomicSavePersistedData(currentData, revision);
-          if (result.success) {
-            persistedData = result.data;
-            console.log(`[Server] Updated lastConnectionDate for team ${team.name} (participant ${userName} joined)`);
-            break;
-          }
-          console.log(`[Server] lastConnectionDate save conflict (attempt ${attempt + 1}), retrying...`);
-        }
-      }
-    } catch (err) {
-      console.warn('[Server] Failed to update lastConnectionDate on session join', err);
-    }
-  });
-
-  // Allow clients to explicitly leave
-  socket.on('leave-session', async () => {
-    await leaveCurrentSession(socket);
-  });
-
-  // Update session data
-  socket.on('update-session', async (sessionData) => {
-    const sessionId = socket.sessionId;
-    if (!sessionId) {
-      console.warn('[Server] update-session received but socket has no sessionId');
-      return;
-    }
-
-    console.log(`[Server] Session update from ${socket.userName}, phase: ${sessionData.phase}`);
-
-    // Store and broadcast to all OTHER clients in the session.
-    // Session updates use last-writer-wins semantics because socket.to()
-    // excludes the sender, so the sender never receives the bumped _rev
-    // and revision-based rejection would permanently block subsequent
-    // updates from the facilitator.
-    try {
-      const savedData = await saveSessionState(sessionId, sessionData);
-      sessions.set(sessionId, savedData);
-
-      // Get room size for logging
-      const room = io.sockets.adapter.rooms.get(sessionId);
-      console.log(`[Server] Broadcasting to ${(room?.size || 1) - 1} other clients in session ${sessionId}`);
-
-      socket.to(sessionId).emit('session-update', savedData);
-    } catch (err) {
-      console.error('[Server] Failed to persist session state', err);
-      // Still try to broadcast even if persist failed
-      sessions.set(sessionId, sessionData);
-      socket.to(sessionId).emit('session-update', sessionData);
-    }
-  });
-
-  // Handle disconnection
-  socket.on('disconnect', async () => {
-    console.log(`[Server] Client disconnected: ${socket.id} (${socket.userName || 'unknown'})`);
-
-    await leaveCurrentSession(socket);
-  });
-});
-
-// Handle SPA routing - serve index.html for all non-API routes
-// Use a regex catch-all compatible with Express 5's path-to-regexp
-app.get(/.*/, (req, res) => {
+app.get(/.*/, (_req, res) => {
   res.sendFile(join(__dirname, 'dist', 'index.html'));
 });
 
 const PORT = process.env.PORT || 3000;
 
-// Initialize database and start server
 const startServer = async () => {
   try {
-    await initDatabase();
-    persistedData = await loadPersistedData();
-    await initSocketAdapter();
+    await dataStore.initDatabase();
+    const persisted = await dataStore.loadPersistedData();
+    dataStore.setPersistedData(persisted);
+    await initSocketAdapter({ io, dataStore });
 
     server.listen(PORT, () => {
       console.log(`[Server] Running on port ${PORT}`);
